Crate harfbuzz_sys

Modules

bindings 🔒

Structs

hb_blob_t

hb_buffer_t

hb_face_t

hb_feature_t

hb_feature_t: @tag: a feature tag @value: 0 disables the feature, non-zero (usually 1) enables the feature. For features implemented as lookup type 3 (like ‘salt’) the @value is a one based index into the alternates. @start: the cluster to start applying this feature setting (inclusive). @end: the cluster to end applying this feature setting (exclusive).

hb_font_extents_t

hb_font_funcs_t

hb_font_t

hb_glyph_extents_t

hb_glyph_info_t

hb_glyph_info_t: @codepoint: either a Unicode code point (before shaping) or a glyph index (after shaping). @cluster: the index of the character in the original text that corresponds to this #hb_glyph_info_t, or whatever the client passes to hb_buffer_add(). More than one #hb_glyph_info_t can have the same @cluster value, if they resulted from the same character (e.g. one to many glyph substitution), and when more than one character gets merged in the same glyph (e.g. many to one glyph substitution) the #hb_glyph_info_t will have the smallest cluster value of them. By default some characters are merged into the same cluster (e.g. combining marks have the same cluster as their bases) even if they are separate glyphs, hb_buffer_set_cluster_level() allow selecting more fine-grained cluster handling.

hb_glyph_position_t

hb_glyph_position_t: @x_advance: how much the line advances after drawing this glyph when setting text in horizontal direction. @y_advance: how much the line advances after drawing this glyph when setting text in vertical direction. @x_offset: how much the glyph moves on the X-axis before drawing it, this should not affect how much the line advances. @y_offset: how much the glyph moves on the Y-axis before drawing it, this should not affect how much the line advances.

hb_language_impl_t

hb_map_t

hb_segment_properties_t

hb_segment_properties_t: @direction: the #hb_direction_t of the buffer, see hb_buffer_set_direction(). @script: the #hb_script_t of the buffer, see hb_buffer_set_script(). @language: the #hb_language_t of the buffer, see hb_buffer_set_language().

hb_set_t

hb_shape_plan_t

hb_unicode_funcs_t

hb_user_data_key_t

hb_variation_t

hb_variation_t:

Constants

HB_BUFFER_CLUSTER_LEVEL_CHARACTERS

HB_BUFFER_CLUSTER_LEVEL_DEFAULT

HB_BUFFER_CLUSTER_LEVEL_MONOTONE_CHARACTERS

HB_BUFFER_CLUSTER_LEVEL_MONOTONE_GRAPHEMES

HB_BUFFER_CONTENT_TYPE_GLYPHS

HB_BUFFER_CONTENT_TYPE_INVALID

HB_BUFFER_CONTENT_TYPE_UNICODE

HB_BUFFER_DIFF_FLAG_CLUSTER_MISMATCH

HB_BUFFER_DIFF_FLAG_CODEPOINT_MISMATCH

HB_BUFFER_DIFF_FLAG_CONTENT_TYPE_MISMATCH

HB_BUFFER_DIFF_FLAG_DOTTED_CIRCLE_PRESENT

HB_BUFFER_DIFF_FLAG_EQUAL

HB_BUFFER_DIFF_FLAG_GLYPH_FLAGS_MISMATCH

HB_BUFFER_DIFF_FLAG_LENGTH_MISMATCH

HB_BUFFER_DIFF_FLAG_NOTDEF_PRESENT

HB_BUFFER_DIFF_FLAG_POSITION_MISMATCH

HB_BUFFER_FLAG_BOT

HB_BUFFER_FLAG_DEFAULT

HB_BUFFER_FLAG_DO_NOT_INSERT_DOTTED_CIRCLE

HB_BUFFER_FLAG_EOT

HB_BUFFER_FLAG_PRESERVE_DEFAULT_IGNORABLES

HB_BUFFER_FLAG_REMOVE_DEFAULT_IGNORABLES

HB_BUFFER_SERIALIZE_FLAG_DEFAULT

HB_BUFFER_SERIALIZE_FLAG_GLYPH_EXTENTS

HB_BUFFER_SERIALIZE_FLAG_GLYPH_FLAGS

HB_BUFFER_SERIALIZE_FLAG_NO_ADVANCES

HB_BUFFER_SERIALIZE_FLAG_NO_CLUSTERS

HB_BUFFER_SERIALIZE_FLAG_NO_GLYPH_NAMES

HB_BUFFER_SERIALIZE_FLAG_NO_POSITIONS

HB_BUFFER_SERIALIZE_FORMAT_INVALID

HB_BUFFER_SERIALIZE_FORMAT_JSON

HB_BUFFER_SERIALIZE_FORMAT_TEXT

HB_DIRECTION_BTT

HB_DIRECTION_INVALID

HB_DIRECTION_LTR

HB_DIRECTION_RTL

HB_DIRECTION_TTB

HB_GLYPH_FLAG_DEFINED

HB_GLYPH_FLAG_UNSAFE_TO_BREAK

HB_MEMORY_MODE_DUPLICATE

HB_MEMORY_MODE_READONLY

HB_MEMORY_MODE_READONLY_MAY_MAKE_WRITABLE

HB_MEMORY_MODE_WRITABLE

HB_SCRIPT_ADLAM

HB_SCRIPT_AHOM

HB_SCRIPT_ANATOLIAN_HIEROGLYPHS

HB_SCRIPT_ARABIC

HB_SCRIPT_ARMENIAN

HB_SCRIPT_AVESTAN

HB_SCRIPT_BALINESE

HB_SCRIPT_BAMUM

HB_SCRIPT_BASSA_VAH

HB_SCRIPT_BATAK

HB_SCRIPT_BENGALI

HB_SCRIPT_BHAIKSUKI

HB_SCRIPT_BOPOMOFO

HB_SCRIPT_BRAHMI

HB_SCRIPT_BRAILLE

HB_SCRIPT_BUGINESE

HB_SCRIPT_BUHID

HB_SCRIPT_CANADIAN_SYLLABICS

HB_SCRIPT_CARIAN

HB_SCRIPT_CAUCASIAN_ALBANIAN

HB_SCRIPT_CHAKMA

HB_SCRIPT_CHAM

HB_SCRIPT_CHEROKEE

HB_SCRIPT_CHORASMIAN

HB_SCRIPT_COMMON

HB_SCRIPT_COPTIC

HB_SCRIPT_CUNEIFORM

HB_SCRIPT_CYPRIOT

HB_SCRIPT_CYRILLIC

HB_SCRIPT_DESERET

HB_SCRIPT_DEVANAGARI

HB_SCRIPT_DIVES_AKURU

HB_SCRIPT_DOGRA

HB_SCRIPT_DUPLOYAN

HB_SCRIPT_EGYPTIAN_HIEROGLYPHS

HB_SCRIPT_ELBASAN

HB_SCRIPT_ELYMAIC

HB_SCRIPT_ETHIOPIC

HB_SCRIPT_GEORGIAN

HB_SCRIPT_GLAGOLITIC

HB_SCRIPT_GOTHIC

HB_SCRIPT_GRANTHA

HB_SCRIPT_GREEK

HB_SCRIPT_GUJARATI

HB_SCRIPT_GUNJALA_GONDI

HB_SCRIPT_GURMUKHI

HB_SCRIPT_HAN

HB_SCRIPT_HANGUL

HB_SCRIPT_HANIFI_ROHINGYA

HB_SCRIPT_HANUNOO

HB_SCRIPT_HATRAN

HB_SCRIPT_HEBREW

HB_SCRIPT_HIRAGANA

HB_SCRIPT_IMPERIAL_ARAMAIC

HB_SCRIPT_INHERITED

HB_SCRIPT_INSCRIPTIONAL_PAHLAVI

HB_SCRIPT_INSCRIPTIONAL_PARTHIAN

HB_SCRIPT_INVALID

HB_SCRIPT_JAVANESE

HB_SCRIPT_KAITHI

HB_SCRIPT_KANNADA

HB_SCRIPT_KATAKANA

HB_SCRIPT_KAYAH_LI

HB_SCRIPT_KHAROSHTHI

HB_SCRIPT_KHITAN_SMALL_SCRIPT

HB_SCRIPT_KHMER

HB_SCRIPT_KHOJKI

HB_SCRIPT_KHUDAWADI

HB_SCRIPT_LAO

HB_SCRIPT_LATIN

HB_SCRIPT_LEPCHA

HB_SCRIPT_LIMBU

HB_SCRIPT_LINEAR_A

HB_SCRIPT_LINEAR_B

HB_SCRIPT_LISU

HB_SCRIPT_LYCIAN

HB_SCRIPT_LYDIAN

HB_SCRIPT_MAHAJANI

HB_SCRIPT_MAKASAR

HB_SCRIPT_MALAYALAM

HB_SCRIPT_MANDAIC

HB_SCRIPT_MANICHAEAN

HB_SCRIPT_MARCHEN

HB_SCRIPT_MASARAM_GONDI

HB_SCRIPT_MEDEFAIDRIN

HB_SCRIPT_MEETEI_MAYEK

HB_SCRIPT_MENDE_KIKAKUI

HB_SCRIPT_MEROITIC_CURSIVE

HB_SCRIPT_MEROITIC_HIEROGLYPHS

HB_SCRIPT_MIAO

HB_SCRIPT_MODI

HB_SCRIPT_MONGOLIAN

HB_SCRIPT_MRO

HB_SCRIPT_MULTANI

HB_SCRIPT_MYANMAR

HB_SCRIPT_NABATAEAN

HB_SCRIPT_NANDINAGARI

HB_SCRIPT_NEWA

HB_SCRIPT_NEW_TAI_LUE

HB_SCRIPT_NKO

HB_SCRIPT_NUSHU

HB_SCRIPT_NYIAKENG_PUACHUE_HMONG

HB_SCRIPT_OGHAM

HB_SCRIPT_OLD_HUNGARIAN

HB_SCRIPT_OLD_ITALIC

HB_SCRIPT_OLD_NORTH_ARABIAN

HB_SCRIPT_OLD_PERMIC

HB_SCRIPT_OLD_PERSIAN

HB_SCRIPT_OLD_SOGDIAN

HB_SCRIPT_OLD_SOUTH_ARABIAN

HB_SCRIPT_OLD_TURKIC

HB_SCRIPT_OL_CHIKI

HB_SCRIPT_ORIYA

HB_SCRIPT_OSAGE

HB_SCRIPT_OSMANYA

HB_SCRIPT_PAHAWH_HMONG

HB_SCRIPT_PALMYRENE

HB_SCRIPT_PAU_CIN_HAU

HB_SCRIPT_PHAGS_PA

HB_SCRIPT_PHOENICIAN

HB_SCRIPT_PSALTER_PAHLAVI

HB_SCRIPT_REJANG

HB_SCRIPT_RUNIC

HB_SCRIPT_SAMARITAN

HB_SCRIPT_SAURASHTRA

HB_SCRIPT_SHARADA

HB_SCRIPT_SHAVIAN

HB_SCRIPT_SIDDHAM

HB_SCRIPT_SIGNWRITING

HB_SCRIPT_SINHALA

HB_SCRIPT_SOGDIAN

HB_SCRIPT_SORA_SOMPENG

HB_SCRIPT_SOYOMBO

HB_SCRIPT_SUNDANESE

HB_SCRIPT_SYLOTI_NAGRI

HB_SCRIPT_SYRIAC

HB_SCRIPT_TAGALOG

HB_SCRIPT_TAGBANWA

HB_SCRIPT_TAI_LE

HB_SCRIPT_TAI_THAM

HB_SCRIPT_TAI_VIET

HB_SCRIPT_TAKRI

HB_SCRIPT_TAMIL

HB_SCRIPT_TANGUT

HB_SCRIPT_TELUGU

HB_SCRIPT_THAANA

HB_SCRIPT_THAI

HB_SCRIPT_TIBETAN

HB_SCRIPT_TIFINAGH

HB_SCRIPT_TIRHUTA

HB_SCRIPT_UGARITIC

HB_SCRIPT_UNKNOWN

HB_SCRIPT_VAI

HB_SCRIPT_WANCHO

HB_SCRIPT_WARANG_CITI

HB_SCRIPT_YEZIDI

HB_SCRIPT_YI

HB_SCRIPT_ZANABAZAR_SQUARE

HB_UNICODE_COMBINING_CLASS_ABOVE

HB_UNICODE_COMBINING_CLASS_ABOVE_LEFT

HB_UNICODE_COMBINING_CLASS_ABOVE_RIGHT

HB_UNICODE_COMBINING_CLASS_ATTACHED_ABOVE

HB_UNICODE_COMBINING_CLASS_ATTACHED_ABOVE_RIGHT

HB_UNICODE_COMBINING_CLASS_ATTACHED_BELOW

HB_UNICODE_COMBINING_CLASS_ATTACHED_BELOW_LEFT

HB_UNICODE_COMBINING_CLASS_BELOW

HB_UNICODE_COMBINING_CLASS_BELOW_LEFT

HB_UNICODE_COMBINING_CLASS_BELOW_RIGHT

HB_UNICODE_COMBINING_CLASS_CCC10

HB_UNICODE_COMBINING_CLASS_CCC11

HB_UNICODE_COMBINING_CLASS_CCC12

HB_UNICODE_COMBINING_CLASS_CCC13

HB_UNICODE_COMBINING_CLASS_CCC14

HB_UNICODE_COMBINING_CLASS_CCC15

HB_UNICODE_COMBINING_CLASS_CCC16

HB_UNICODE_COMBINING_CLASS_CCC17

HB_UNICODE_COMBINING_CLASS_CCC18

HB_UNICODE_COMBINING_CLASS_CCC19

HB_UNICODE_COMBINING_CLASS_CCC20

HB_UNICODE_COMBINING_CLASS_CCC21

HB_UNICODE_COMBINING_CLASS_CCC22

HB_UNICODE_COMBINING_CLASS_CCC23

HB_UNICODE_COMBINING_CLASS_CCC24

HB_UNICODE_COMBINING_CLASS_CCC25

HB_UNICODE_COMBINING_CLASS_CCC26

HB_UNICODE_COMBINING_CLASS_CCC27

HB_UNICODE_COMBINING_CLASS_CCC28

HB_UNICODE_COMBINING_CLASS_CCC29

HB_UNICODE_COMBINING_CLASS_CCC30

HB_UNICODE_COMBINING_CLASS_CCC31

HB_UNICODE_COMBINING_CLASS_CCC32

HB_UNICODE_COMBINING_CLASS_CCC33

HB_UNICODE_COMBINING_CLASS_CCC34

HB_UNICODE_COMBINING_CLASS_CCC35

HB_UNICODE_COMBINING_CLASS_CCC36

HB_UNICODE_COMBINING_CLASS_CCC84

HB_UNICODE_COMBINING_CLASS_CCC91

HB_UNICODE_COMBINING_CLASS_CCC103

HB_UNICODE_COMBINING_CLASS_CCC107

HB_UNICODE_COMBINING_CLASS_CCC118

HB_UNICODE_COMBINING_CLASS_CCC122

HB_UNICODE_COMBINING_CLASS_CCC129

HB_UNICODE_COMBINING_CLASS_CCC130

HB_UNICODE_COMBINING_CLASS_CCC133

HB_UNICODE_COMBINING_CLASS_DOUBLE_ABOVE

HB_UNICODE_COMBINING_CLASS_DOUBLE_BELOW

HB_UNICODE_COMBINING_CLASS_INVALID

HB_UNICODE_COMBINING_CLASS_IOTA_SUBSCRIPT

HB_UNICODE_COMBINING_CLASS_KANA_VOICING

HB_UNICODE_COMBINING_CLASS_LEFT

HB_UNICODE_COMBINING_CLASS_NOT_REORDERED

HB_UNICODE_COMBINING_CLASS_NUKTA

HB_UNICODE_COMBINING_CLASS_OVERLAY

HB_UNICODE_COMBINING_CLASS_RIGHT

HB_UNICODE_COMBINING_CLASS_VIRAMA

HB_UNICODE_GENERAL_CATEGORY_CLOSE_PUNCTUATION

HB_UNICODE_GENERAL_CATEGORY_CONNECT_PUNCTUATION

HB_UNICODE_GENERAL_CATEGORY_CONTROL

HB_UNICODE_GENERAL_CATEGORY_CURRENCY_SYMBOL

HB_UNICODE_GENERAL_CATEGORY_DASH_PUNCTUATION

HB_UNICODE_GENERAL_CATEGORY_DECIMAL_NUMBER

HB_UNICODE_GENERAL_CATEGORY_ENCLOSING_MARK

HB_UNICODE_GENERAL_CATEGORY_FINAL_PUNCTUATION

HB_UNICODE_GENERAL_CATEGORY_FORMAT

HB_UNICODE_GENERAL_CATEGORY_INITIAL_PUNCTUATION

HB_UNICODE_GENERAL_CATEGORY_LETTER_NUMBER

HB_UNICODE_GENERAL_CATEGORY_LINE_SEPARATOR

HB_UNICODE_GENERAL_CATEGORY_LOWERCASE_LETTER

HB_UNICODE_GENERAL_CATEGORY_MATH_SYMBOL

HB_UNICODE_GENERAL_CATEGORY_MODIFIER_LETTER

HB_UNICODE_GENERAL_CATEGORY_MODIFIER_SYMBOL

HB_UNICODE_GENERAL_CATEGORY_NON_SPACING_MARK

HB_UNICODE_GENERAL_CATEGORY_OPEN_PUNCTUATION

HB_UNICODE_GENERAL_CATEGORY_OTHER_LETTER

HB_UNICODE_GENERAL_CATEGORY_OTHER_NUMBER

HB_UNICODE_GENERAL_CATEGORY_OTHER_PUNCTUATION

HB_UNICODE_GENERAL_CATEGORY_OTHER_SYMBOL

HB_UNICODE_GENERAL_CATEGORY_PARAGRAPH_SEPARATOR

HB_UNICODE_GENERAL_CATEGORY_PRIVATE_USE

HB_UNICODE_GENERAL_CATEGORY_SPACE_SEPARATOR

HB_UNICODE_GENERAL_CATEGORY_SPACING_MARK

HB_UNICODE_GENERAL_CATEGORY_SURROGATE

HB_UNICODE_GENERAL_CATEGORY_TITLECASE_LETTER

HB_UNICODE_GENERAL_CATEGORY_UNASSIGNED

HB_UNICODE_GENERAL_CATEGORY_UPPERCASE_LETTER

_HB_SCRIPT_MAX_VALUE

_HB_SCRIPT_MAX_VALUE_SIGNED

Functions

hb_blob_copy_writable_or_fail⚠

hb_blob_create⚠

hb_blob_create_from_file⚠

hb_blob_create_sub_blob⚠

hb_blob_destroy⚠

hb_blob_get_data⚠

hb_blob_get_data_writable⚠

hb_blob_get_empty⚠

hb_blob_get_length⚠

hb_blob_get_user_data⚠

hb_blob_is_immutable⚠

hb_blob_make_immutable⚠

hb_blob_reference⚠

hb_blob_set_user_data⚠

hb_buffer_add⚠

hb_buffer_add_codepoints⚠

hb_buffer_add_latin1⚠

hb_buffer_add_utf8⚠

hb_buffer_add_utf16⚠

hb_buffer_add_utf32⚠

hb_buffer_allocation_successful⚠

hb_buffer_append⚠

hb_buffer_clear_contents⚠

hb_buffer_create⚠

hb_buffer_deserialize_glyphs⚠

hb_buffer_destroy⚠

hb_buffer_diff⚠

hb_buffer_get_cluster_level⚠

hb_buffer_get_content_type⚠

hb_buffer_get_direction⚠

hb_buffer_get_empty⚠

hb_buffer_get_flags⚠

hb_buffer_get_glyph_infos⚠

hb_buffer_get_glyph_positions⚠

hb_buffer_get_invisible_glyph⚠

hb_buffer_get_language⚠

hb_buffer_get_length⚠

hb_buffer_get_replacement_codepoint⚠

hb_buffer_get_script⚠

hb_buffer_get_segment_properties⚠

hb_buffer_get_unicode_funcs⚠

hb_buffer_get_user_data⚠

hb_buffer_guess_segment_properties⚠

hb_buffer_normalize_glyphs⚠

hb_buffer_pre_allocate⚠

hb_buffer_reference⚠

hb_buffer_reset⚠

hb_buffer_reverse⚠

hb_buffer_reverse_clusters⚠

hb_buffer_reverse_range⚠

hb_buffer_serialize_format_from_string⚠

hb_buffer_serialize_format_to_string⚠

hb_buffer_serialize_glyphs⚠

hb_buffer_serialize_list_formats⚠

hb_buffer_set_cluster_level⚠

hb_buffer_set_content_type⚠

hb_buffer_set_direction⚠

hb_buffer_set_flags⚠

hb_buffer_set_invisible_glyph⚠

hb_buffer_set_language⚠

hb_buffer_set_length⚠

hb_buffer_set_message_func⚠

hb_buffer_set_replacement_codepoint⚠

hb_buffer_set_script⚠

hb_buffer_set_segment_properties⚠

hb_buffer_set_unicode_funcs⚠

hb_buffer_set_user_data⚠

hb_color_get_alpha⚠

hb_color_get_blue⚠

hb_color_get_green⚠

hb_color_get_red⚠

hb_direction_from_string⚠

hb_direction_to_string⚠

hb_face_builder_add_table⚠

hb_face_builder_create⚠

hb_face_collect_unicodes⚠

hb_face_collect_variation_selectors⚠

hb_face_collect_variation_unicodes⚠

hb_face_count⚠

hb_face_create⚠

hb_face_create_for_tables⚠

hb_face_destroy⚠

hb_face_get_empty⚠

hb_face_get_glyph_count⚠

hb_face_get_index⚠

hb_face_get_table_tags⚠

hb_face_get_upem⚠

hb_face_get_user_data⚠

hb_face_is_immutable⚠

hb_face_make_immutable⚠

hb_face_reference⚠

hb_face_reference_blob⚠

hb_face_reference_table⚠

hb_face_set_glyph_count⚠

hb_face_set_index⚠

hb_face_set_upem⚠

hb_face_set_user_data⚠

hb_feature_from_string⚠

hb_feature_to_string⚠

hb_font_add_glyph_origin_for_direction⚠

hb_font_create⚠

hb_font_create_sub_font⚠

hb_font_destroy⚠

hb_font_funcs_create⚠

hb_font_funcs_destroy⚠

hb_font_funcs_get_empty⚠

hb_font_funcs_get_user_data⚠

hb_font_funcs_is_immutable⚠

hb_font_funcs_make_immutable⚠

hb_font_funcs_reference⚠

hb_font_funcs_set_font_h_extents_func⚠

hb_font_funcs_set_font_h_extents_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_font_v_extents_func⚠

hb_font_funcs_set_font_v_extents_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_contour_point_func⚠

hb_font_funcs_set_glyph_contour_point_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_extents_func⚠

hb_font_funcs_set_glyph_extents_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_from_name_func⚠

hb_font_funcs_set_glyph_from_name_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_func⚠

hb_font_funcs_set_glyph_h_advance_func⚠

hb_font_funcs_set_glyph_h_advance_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_h_advances_func⚠

hb_font_funcs_set_glyph_h_advances_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_h_kerning_func⚠

hb_font_funcs_set_glyph_h_kerning_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_h_origin_func⚠

hb_font_funcs_set_glyph_h_origin_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_name_func⚠

hb_font_funcs_set_glyph_name_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_v_advance_func⚠

hb_font_funcs_set_glyph_v_advance_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_v_advances_func⚠

hb_font_funcs_set_glyph_v_advances_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_v_kerning_func⚠

hb_font_funcs_set_glyph_v_kerning_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_v_origin_func⚠

hb_font_funcs_set_glyph_v_origin_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_nominal_glyph_func⚠

hb_font_funcs_set_nominal_glyph_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_nominal_glyphs_func⚠

hb_font_funcs_set_nominal_glyphs_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_user_data⚠

hb_font_funcs_set_variation_glyph_func⚠

hb_font_funcs_set_variation_glyph_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_get_empty⚠

hb_font_get_extents_for_direction⚠

hb_font_get_face⚠

hb_font_get_glyph⚠

hb_font_get_glyph_advance_for_direction⚠

hb_font_get_glyph_advances_for_direction⚠

hb_font_get_glyph_contour_point⚠

hb_font_get_glyph_contour_point_for_origin⚠

hb_font_get_glyph_extents⚠

hb_font_get_glyph_extents_for_origin⚠

hb_font_get_glyph_from_name⚠

hb_font_get_glyph_h_advance⚠

hb_font_get_glyph_h_advances⚠

hb_font_get_glyph_h_kerning⚠

hb_font_get_glyph_h_origin⚠

hb_font_get_glyph_kerning_for_direction⚠

hb_font_get_glyph_name⚠

hb_font_get_glyph_origin_for_direction⚠

hb_font_get_glyph_v_advance⚠

hb_font_get_glyph_v_advances⚠

hb_font_get_glyph_v_kerning⚠

hb_font_get_glyph_v_origin⚠

hb_font_get_h_extents⚠

hb_font_get_nominal_glyph⚠

hb_font_get_nominal_glyphs⚠

hb_font_get_parent⚠

hb_font_get_ppem⚠

hb_font_get_ptem⚠

hb_font_get_scale⚠

hb_font_get_user_data⚠

hb_font_get_v_extents⚠

hb_font_get_var_coords_normalized⚠

hb_font_get_variation_glyph⚠

hb_font_glyph_from_string⚠

hb_font_glyph_to_string⚠

hb_font_is_immutable⚠

hb_font_make_immutable⚠

hb_font_reference⚠

hb_font_set_face⚠

hb_font_set_funcs⚠

hb_font_set_funcs_data⚠

hb_font_set_parent⚠

hb_font_set_ppem⚠

hb_font_set_ptem⚠

hb_font_set_scale⚠

hb_font_set_user_data⚠

hb_font_set_var_coords_design⚠

hb_font_set_var_coords_normalized⚠

hb_font_set_var_named_instance⚠

hb_font_set_variations⚠

hb_font_subtract_glyph_origin_for_direction⚠

hb_ft_font_create_referenced⚠

hb_glyph_info_get_glyph_flags⚠

hb_language_from_string⚠

hb_language_get_default⚠

hb_language_to_string⚠

hb_map_allocation_successful⚠

hb_map_clear⚠

hb_map_create⚠

hb_map_del⚠

hb_map_destroy⚠

hb_map_get⚠

hb_map_get_empty⚠

hb_map_get_population⚠

hb_map_get_user_data⚠

hb_map_has⚠

hb_map_is_empty⚠

hb_map_reference⚠

hb_map_set⚠

hb_map_set_user_data⚠

hb_script_from_iso15924_tag⚠

hb_script_from_string⚠

hb_script_get_horizontal_direction⚠

hb_script_to_iso15924_tag⚠

hb_segment_properties_equal⚠

hb_segment_properties_hash⚠

hb_set_add⚠

hb_set_add_range⚠

hb_set_allocation_successful⚠

hb_set_clear⚠

hb_set_create⚠

hb_set_del⚠

hb_set_del_range⚠

hb_set_destroy⚠

hb_set_get_empty⚠

hb_set_get_max⚠

hb_set_get_min⚠

hb_set_get_population⚠

hb_set_get_user_data⚠

hb_set_has⚠

hb_set_intersect⚠

hb_set_invert⚠

hb_set_is_empty⚠

hb_set_is_equal⚠

hb_set_is_subset⚠

hb_set_next⚠

hb_set_next_range⚠

hb_set_previous⚠

hb_set_previous_range⚠

hb_set_reference⚠

hb_set_set⚠

hb_set_set_user_data⚠

hb_set_subtract⚠

hb_set_symmetric_difference⚠

hb_set_union⚠

hb_shape⚠

hb_shape_full⚠

hb_shape_list_shapers⚠

hb_shape_plan_create⚠

hb_shape_plan_create2⚠

hb_shape_plan_create_cached⚠

hb_shape_plan_create_cached2⚠

hb_shape_plan_destroy⚠

hb_shape_plan_execute⚠

hb_shape_plan_get_empty⚠

hb_shape_plan_get_shaper⚠

hb_shape_plan_get_user_data⚠

hb_shape_plan_reference⚠

hb_shape_plan_set_user_data⚠

hb_tag_from_string⚠

hb_tag_to_string⚠

hb_unicode_combining_class⚠

hb_unicode_combining_class:

hb_unicode_compose⚠

hb_unicode_decompose⚠

hb_unicode_decompose_compatibility⚠

hb_unicode_eastasian_width⚠

hb_unicode_eastasian_width:

hb_unicode_funcs_create⚠

hb_unicode_funcs_destroy⚠

hb_unicode_funcs_get_default⚠

hb_unicode_funcs_get_empty⚠

hb_unicode_funcs_get_parent⚠

hb_unicode_funcs_get_user_data⚠

hb_unicode_funcs_is_immutable⚠

hb_unicode_funcs_make_immutable⚠

hb_unicode_funcs_reference⚠

hb_unicode_funcs_set_combining_class_func⚠

hb_unicode_funcs_set_combining_class_func: @ufuncs: a Unicode function structure @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_unicode_funcs_set_compose_func⚠

hb_unicode_funcs_set_compose_func: @ufuncs: a Unicode function structure @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_unicode_funcs_set_decompose_compatibility_func⚠

hb_unicode_funcs_set_decompose_compatibility_func: @ufuncs: a Unicode function structure @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_unicode_funcs_set_decompose_func⚠

hb_unicode_funcs_set_decompose_func: @ufuncs: a Unicode function structure @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_unicode_funcs_set_eastasian_width_func⚠

hb_unicode_funcs_set_eastasian_width_func: @ufuncs: a Unicode function structure @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_unicode_funcs_set_general_category_func⚠

hb_unicode_funcs_set_general_category_func: @ufuncs: a Unicode function structure @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_unicode_funcs_set_mirroring_func⚠

hb_unicode_funcs_set_mirroring_func: @ufuncs: a Unicode function structure @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_unicode_funcs_set_script_func⚠

hb_unicode_funcs_set_script_func: @ufuncs: a Unicode function structure @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_unicode_funcs_set_user_data⚠

hb_unicode_general_category⚠

hb_unicode_general_category:

hb_unicode_mirroring⚠

hb_unicode_mirroring:

hb_unicode_script⚠

hb_unicode_script:

hb_variation_from_string⚠

hb_variation_to_string⚠

hb_version⚠

hb_version_atleast⚠

hb_version_string⚠

Type Aliases

hb_bool_t

hb_buffer_cluster_level_t

hb_buffer_cluster_level_t: @HB_BUFFER_CLUSTER_LEVEL_MONOTONE_GRAPHEMES: Return cluster values grouped by graphemes into monotone order. @HB_BUFFER_CLUSTER_LEVEL_MONOTONE_CHARACTERS: Return cluster values grouped into monotone order. @HB_BUFFER_CLUSTER_LEVEL_CHARACTERS: Don’t group cluster values. @HB_BUFFER_CLUSTER_LEVEL_DEFAULT: Default cluster level, equal to @HB_BUFFER_CLUSTER_LEVEL_MONOTONE_GRAPHEMES.

hb_buffer_content_type_t

hb_buffer_content_type_t: @HB_BUFFER_CONTENT_TYPE_INVALID: Initial value for new buffer. @HB_BUFFER_CONTENT_TYPE_UNICODE: The buffer contains input characters (before shaping). @HB_BUFFER_CONTENT_TYPE_GLYPHS: The buffer contains output glyphs (after shaping).

hb_buffer_diff_flags_t

hb_buffer_flags_t

hb_buffer_flags_t: @HB_BUFFER_FLAG_DEFAULT: the default buffer flag. @HB_BUFFER_FLAG_BOT: flag indicating that special handling of the beginning of text paragraph can be applied to this buffer. Should usually be set, unless you are passing to the buffer only part of the text without the full context. @HB_BUFFER_FLAG_EOT: flag indicating that special handling of the end of text paragraph can be applied to this buffer, similar to @HB_BUFFER_FLAG_BOT. @HB_BUFFER_FLAG_PRESERVE_DEFAULT_IGNORABLES: flag indication that character with Default_Ignorable Unicode property should use the corresponding glyph from the font, instead of hiding them (done by replacing them with the space glyph and zeroing the advance width.) This flag takes precedence over @HB_BUFFER_FLAG_REMOVE_DEFAULT_IGNORABLES. @HB_BUFFER_FLAG_REMOVE_DEFAULT_IGNORABLES: flag indication that character with Default_Ignorable Unicode property should be removed from glyph string instead of hiding them (done by replacing them with the space glyph and zeroing the advance width.) @HB_BUFFER_FLAG_PRESERVE_DEFAULT_IGNORABLES takes precedence over this flag. Since: 1.8.0 @HB_BUFFER_FLAG_DO_NOT_INSERT_DOTTED_CIRCLE: flag indicating that a dotted circle should not be inserted in the rendering of incorrect character sequences (such at <0905 093E>). Since: 2.4

hb_buffer_message_func_t

hb_buffer_serialize_flags_t

hb_buffer_serialize_flags_t: @HB_BUFFER_SERIALIZE_FLAG_DEFAULT: serialize glyph names, clusters and positions. @HB_BUFFER_SERIALIZE_FLAG_NO_CLUSTERS: do not serialize glyph cluster. @HB_BUFFER_SERIALIZE_FLAG_NO_POSITIONS: do not serialize glyph position information. @HB_BUFFER_SERIALIZE_FLAG_NO_GLYPH_NAMES: do no serialize glyph name. @HB_BUFFER_SERIALIZE_FLAG_GLYPH_EXTENTS: serialize glyph extents. @HB_BUFFER_SERIALIZE_FLAG_GLYPH_FLAGS: serialize glyph flags. Since: 1.5.0 @HB_BUFFER_SERIALIZE_FLAG_NO_ADVANCES: do not serialize glyph advances, glyph offsets will reflect absolute glyph positions. Since: 1.8.0

hb_buffer_serialize_format_t

hb_buffer_serialize_format_t: @HB_BUFFER_SERIALIZE_FORMAT_TEXT: a human-readable, plain text format. @HB_BUFFER_SERIALIZE_FORMAT_JSON: a machine-readable JSON format. @HB_BUFFER_SERIALIZE_FORMAT_INVALID: invalid format.

hb_codepoint_t

hb_color_t

hb_color_t:

hb_destroy_func_t

hb_direction_t

hb_direction_t: @HB_DIRECTION_INVALID: Initial, unset direction. @HB_DIRECTION_LTR: Text is set horizontally from left to right. @HB_DIRECTION_RTL: Text is set horizontally from right to left. @HB_DIRECTION_TTB: Text is set vertically from top to bottom. @HB_DIRECTION_BTT: Text is set vertically from bottom to top.

hb_font_get_font_extents_func_t

hb_font_get_font_h_extents_func_t

hb_font_get_font_v_extents_func_t

hb_font_get_glyph_advance_func_t

hb_font_get_glyph_advances_func_t

hb_font_get_glyph_contour_point_func_t

hb_font_get_glyph_extents_func_t

hb_font_get_glyph_from_name_func_t

hb_font_get_glyph_func_t

hb_font_get_glyph_h_advance_func_t

hb_font_get_glyph_h_advances_func_t

hb_font_get_glyph_h_kerning_func_t

hb_font_get_glyph_h_origin_func_t

hb_font_get_glyph_kerning_func_t

hb_font_get_glyph_name_func_t

hb_font_get_glyph_origin_func_t

hb_font_get_glyph_v_advance_func_t

hb_font_get_glyph_v_advances_func_t

hb_font_get_glyph_v_kerning_func_t

hb_font_get_glyph_v_origin_func_t

hb_font_get_nominal_glyph_func_t

hb_font_get_nominal_glyphs_func_t

hb_font_get_variation_glyph_func_t

hb_glyph_flags_t

hb_glyph_flags_t: @HB_GLYPH_FLAG_UNSAFE_TO_BREAK: Indicates that if input text is broken at the beginning of the cluster this glyph is part of, then both sides need to be re-shaped, as the result might be different. On the flip side, it means that when this flag is not present, then it’s safe to break the glyph-run at the beginning of this cluster, and the two sides represent the exact same result one would get if breaking input text at the beginning of this cluster and shaping the two sides separately. This can be used to optimize paragraph layout, by avoiding re-shaping of each line after line-breaking, or limiting the reshaping to a small piece around the breaking point only. @HB_GLYPH_FLAG_DEFINED: All the currently defined flags.

hb_language_t

hb_mask_t

hb_memory_mode_t

hb_position_t

hb_reference_table_func_t

hb_script_t

hb_tag_t

hb_unicode_combining_class_func_t

hb_unicode_combining_class_t

hb_unicode_compose_func_t

hb_unicode_decompose_compatibility_func_t

hb_unicode_decompose_compatibility_func_t: @ufuncs: a Unicode function structure @u: codepoint to decompose @decomposed: address of codepoint array (of length %HB_UNICODE_MAX_DECOMPOSITION_LEN) to write decomposition into @user_data: user data pointer as passed to hb_unicode_funcs_set_decompose_compatibility_func()

hb_unicode_decompose_func_t

hb_unicode_eastasian_width_func_t

hb_unicode_eastasian_width_func_t:

hb_unicode_general_category_func_t

hb_unicode_general_category_t

hb_unicode_mirroring_func_t

hb_unicode_script_func_t

hb_var_int_t

Unions

_hb_var_int_t