harfbuzz_sys

Type Alias hb_ot_color_palette_flags_t

pub type hb_ot_color_palette_flags_t = c_uint;

hb_ot_color_palette_flags_t: @HB_OT_COLOR_PALETTE_FLAG_DEFAULT: Default indicating that there is nothing special to note about a color palette. @HB_OT_COLOR_PALETTE_FLAG_USABLE_WITH_LIGHT_BACKGROUND: Flag indicating that the color palette is appropriate to use when displaying the font on a light background such as white. @HB_OT_COLOR_PALETTE_FLAG_USABLE_WITH_DARK_BACKGROUND: Flag indicating that the color palette is appropriate to use when displaying the font on a dark background such as black.

Since: 2.1.0