harfbuzz_sys

Type Alias hb_unicode_combining_class_t

pub type hb_unicode_combining_class_t = c_uint;

hb_unicode_combining_class_t: @HB_UNICODE_COMBINING_CLASS_NOT_REORDERED: Spacing and enclosing marks; also many vowel and consonant signs, even if nonspacing @HB_UNICODE_COMBINING_CLASS_OVERLAY: Marks which overlay a base letter or symbol @HB_UNICODE_COMBINING_CLASS_NUKTA: Diacritic nukta marks in Brahmi-derived scripts @HB_UNICODE_COMBINING_CLASS_KANA_VOICING: Hiragana/Katakana voicing marks @HB_UNICODE_COMBINING_CLASS_VIRAMA: Viramas @HB_UNICODE_COMBINING_CLASS_CCC10: [Hebrew] @HB_UNICODE_COMBINING_CLASS_CCC11: [Hebrew] @HB_UNICODE_COMBINING_CLASS_CCC12: [Hebrew] @HB_UNICODE_COMBINING_CLASS_CCC13: [Hebrew] @HB_UNICODE_COMBINING_CLASS_CCC14: [Hebrew] @HB_UNICODE_COMBINING_CLASS_CCC15: [Hebrew] @HB_UNICODE_COMBINING_CLASS_CCC16: [Hebrew] @HB_UNICODE_COMBINING_CLASS_CCC17: [Hebrew] @HB_UNICODE_COMBINING_CLASS_CCC18: [Hebrew] @HB_UNICODE_COMBINING_CLASS_CCC19: [Hebrew] @HB_UNICODE_COMBINING_CLASS_CCC20: [Hebrew] @HB_UNICODE_COMBINING_CLASS_CCC21: [Hebrew] @HB_UNICODE_COMBINING_CLASS_CCC22: [Hebrew] @HB_UNICODE_COMBINING_CLASS_CCC23: [Hebrew] @HB_UNICODE_COMBINING_CLASS_CCC24: [Hebrew] @HB_UNICODE_COMBINING_CLASS_CCC25: [Hebrew] @HB_UNICODE_COMBINING_CLASS_CCC26: [Hebrew] @HB_UNICODE_COMBINING_CLASS_CCC27: [Arabic] @HB_UNICODE_COMBINING_CLASS_CCC28: [Arabic] @HB_UNICODE_COMBINING_CLASS_CCC29: [Arabic] @HB_UNICODE_COMBINING_CLASS_CCC30: [Arabic] @HB_UNICODE_COMBINING_CLASS_CCC31: [Arabic] @HB_UNICODE_COMBINING_CLASS_CCC32: [Arabic] @HB_UNICODE_COMBINING_CLASS_CCC33: [Arabic] @HB_UNICODE_COMBINING_CLASS_CCC34: [Arabic] @HB_UNICODE_COMBINING_CLASS_CCC35: [Arabic] @HB_UNICODE_COMBINING_CLASS_CCC36: [Syriac] @HB_UNICODE_COMBINING_CLASS_CCC84: [Telugu] @HB_UNICODE_COMBINING_CLASS_CCC91: [Telugu] @HB_UNICODE_COMBINING_CLASS_CCC103: [Thai] @HB_UNICODE_COMBINING_CLASS_CCC107: [Thai] @HB_UNICODE_COMBINING_CLASS_CCC118: [Lao] @HB_UNICODE_COMBINING_CLASS_CCC122: [Lao] @HB_UNICODE_COMBINING_CLASS_CCC129: [Tibetan] @HB_UNICODE_COMBINING_CLASS_CCC130: [Tibetan] @HB_UNICODE_COMBINING_CLASS_CCC133: [Tibetan] @HB_UNICODE_COMBINING_CLASS_ATTACHED_BELOW_LEFT: Marks attached at the bottom left @HB_UNICODE_COMBINING_CLASS_ATTACHED_BELOW: Marks attached directly below @HB_UNICODE_COMBINING_CLASS_ATTACHED_ABOVE: Marks attached directly above @HB_UNICODE_COMBINING_CLASS_ATTACHED_ABOVE_RIGHT: Marks attached at the top right @HB_UNICODE_COMBINING_CLASS_BELOW_LEFT: Distinct marks at the bottom left @HB_UNICODE_COMBINING_CLASS_BELOW: Distinct marks directly below @HB_UNICODE_COMBINING_CLASS_BELOW_RIGHT: Distinct marks at the bottom right @HB_UNICODE_COMBINING_CLASS_LEFT: Distinct marks to the left @HB_UNICODE_COMBINING_CLASS_RIGHT: Distinct marks to the right @HB_UNICODE_COMBINING_CLASS_ABOVE_LEFT: Distinct marks at the top left @HB_UNICODE_COMBINING_CLASS_ABOVE: Distinct marks directly above @HB_UNICODE_COMBINING_CLASS_ABOVE_RIGHT: Distinct marks at the top right @HB_UNICODE_COMBINING_CLASS_DOUBLE_BELOW: Distinct marks subtending two bases @HB_UNICODE_COMBINING_CLASS_DOUBLE_ABOVE: Distinct marks extending above two bases @HB_UNICODE_COMBINING_CLASS_IOTA_SUBSCRIPT: Greek iota subscript only @HB_UNICODE_COMBINING_CLASS_INVALID: Invalid combining class

Data type for the Canonical_Combining_Class (ccc) property from the Unicode Character Database.

Note: newer versions of Unicode may add new values. Client programs should be ready to handle any value in the 0..254 range being returned from hb_unicode_combining_class().