Type Alias harfbuzz_sys::hb_unicode_compose_func_t

pub type hb_unicode_compose_func_t = Option<unsafe extern "C" fn(ufuncs: *mut hb_unicode_funcs_t, a: hb_codepoint_t, b: hb_codepoint_t, ab: *mut hb_codepoint_t, user_data: *mut c_void) -> hb_bool_t>;

hb_unicode_compose_func_t: @ufuncs: A Unicode-functions structure @a: The first code point to compose @b: The second code point to compose @ab: (out): The composed code point @user_data: user data pointer passed by the caller

A virtual method for the #hb_unicode_funcs_t structure.

This method should compose a sequence of two input Unicode code points by canonical equivalence, returning the composed code point in a #hb_codepoint_t output parameter (if successful). The method must return an #hb_bool_t indicating the success of the composition.

Return value: True is @a,@b composed, false otherwise

Aliased Type

enum hb_unicode_compose_func_t {
    None,
    Some(unsafe extern "C" fn(_: *mut hb_unicode_funcs_t, _: u32, _: u32, _: *mut u32, _: *mut c_void) -> i32),
}

Variants

None

No value.

Some(unsafe extern "C" fn(_: *mut hb_unicode_funcs_t, _: u32, _: u32, _: *mut u32, _: *mut c_void) -> i32)

Some value of type T.