Crate harfbuzz_sys

Structs

hb_aat_layout_feature_selector_info_t

hb_blob_t

hb_buffer_t

hb_face_t

hb_feature_t

hb_font_extents_t

hb_font_funcs_t

hb_font_t

hb_glyph_extents_t

hb_glyph_info_t
hb_glyph_info_t: @codepoint: either a Unicode code point (before shaping) or a glyph index (after shaping). @mask: @cluster: the index of the character in the original text that corresponds to this hb_glyph_info_t, or whatever the client passes to hb_buffer_add(). More than one hb_glyph_info_t can have the same @cluster value, if they resulted from the same character (e.g. one to many glyph substitution), and when more than one character gets merged in the same glyph (e.g. many to one glyph substitution) the hb_glyph_info_t will have the smallest cluster value of them. By default some characters are merged into the same cluster (e.g. combining marks have the same cluster as their bases) even if they are separate glyphs, hb_buffer_set_cluster_level() allow selecting more fine-grained cluster handling.

hb_glyph_position_t
hb_glyph_position_t: @x_advance: how much the line advances after drawing this glyph when setting text in horizontal direction. @y_advance: how much the line advances after drawing this glyph when setting text in vertical direction. @x_offset: how much the glyph moves on the X-axis before drawing it, this should not affect how much the line advances. @y_offset: how much the glyph moves on the Y-axis before drawing it, this should not affect how much the line advances.

hb_language_impl_t

hb_map_t

hb_ot_color_layer_t
hb_ot_color_layer_t:

hb_ot_math_glyph_part_t
hb_ot_math_glyph_part_t: @glyph: The glyph index of the variant part @start_connector_length: The length of the connector on the starting side of the variant part @end_connector_length: The length of the connector on the ending side of the variant part @full_advance: The total advance of the part @flags: hb_ot_math_glyph_part_flags_t flags for the part

hb_ot_math_glyph_variant_t
hb_ot_math_glyph_variant_t: @glyph: The glyph index of the variant @advance: The advance width of the variant

hb_ot_name_entry_t
hb_ot_name_entry_t: @name_id: name ID @language: language

hb_ot_var_axis_info_t
hb_ot_var_axis_info_t:

hb_segment_properties_t
hb_segment_properties_t: @direction: the hb_direction_t of the buffer, see hb_buffer_set_direction(). @script: the hb_script_t of the buffer, see hb_buffer_set_script(). @language: the hb_language_t of the buffer, see hb_buffer_set_language().

hb_set_t

hb_shape_plan_t

hb_unicode_funcs_t

hb_user_data_key_t

hb_variation_t
hb_variation_t:

Constants

HB_AAT_LAYOUT_FEATURE_SELECTOR_ABBREV_SQUARED_LIGATURES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_ABBREV_SQUARED_LIGATURES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_ALL_CAPS

HB_AAT_LAYOUT_FEATURE_SELECTOR_ALL_LOWER_CASE

HB_AAT_LAYOUT_FEATURE_SELECTOR_ALL_TYPE_FEATURES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_ALL_TYPE_FEATURES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_ALTERNATE_HORIZ_KANA_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_ALTERNATE_HORIZ_KANA_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_ALTERNATE_VERT_KANA_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_ALTERNATE_VERT_KANA_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_ALT_HALF_WIDTH_TEXT

HB_AAT_LAYOUT_FEATURE_SELECTOR_ALT_PROPORTIONAL_TEXT

HB_AAT_LAYOUT_FEATURE_SELECTOR_ASTERISK_TO_MULTIPLY_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_ASTERISK_TO_MULTIPLY_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_BOX_ANNOTATION

HB_AAT_LAYOUT_FEATURE_SELECTOR_CANONICAL_COMPOSITION_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_CANONICAL_COMPOSITION_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_CASE_SENSITIVE_LAYOUT_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_CASE_SENSITIVE_LAYOUT_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_CASE_SENSITIVE_SPACING_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_CASE_SENSITIVE_SPACING_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_CIRCLE_ANNOTATION

HB_AAT_LAYOUT_FEATURE_SELECTOR_CJK_ITALIC_ROMAN

HB_AAT_LAYOUT_FEATURE_SELECTOR_CJK_ITALIC_ROMAN_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_CJK_ITALIC_ROMAN_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_CJK_SYMBOL_ALT_FIVE

HB_AAT_LAYOUT_FEATURE_SELECTOR_CJK_SYMBOL_ALT_FOUR

HB_AAT_LAYOUT_FEATURE_SELECTOR_CJK_SYMBOL_ALT_ONE

HB_AAT_LAYOUT_FEATURE_SELECTOR_CJK_SYMBOL_ALT_THREE

HB_AAT_LAYOUT_FEATURE_SELECTOR_CJK_SYMBOL_ALT_TWO

HB_AAT_LAYOUT_FEATURE_SELECTOR_CJK_VERTICAL_ROMAN_CENTERED

HB_AAT_LAYOUT_FEATURE_SELECTOR_CJK_VERTICAL_ROMAN_HBASELINE

HB_AAT_LAYOUT_FEATURE_SELECTOR_COMMON_LIGATURES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_COMMON_LIGATURES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_COMPATIBILITY_COMPOSITION_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_COMPATIBILITY_COMPOSITION_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_CONTEXTUAL_ALTERNATES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_CONTEXTUAL_ALTERNATES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_CONTEXTUAL_LIGATURES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_CONTEXTUAL_LIGATURES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_CONTEXTUAL_SWASH_ALTERNATES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_CONTEXTUAL_SWASH_ALTERNATES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_CURSIVE

HB_AAT_LAYOUT_FEATURE_SELECTOR_DECOMPOSE_DIACRITICS

HB_AAT_LAYOUT_FEATURE_SELECTOR_DECORATIVE_BORDERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_DEFAULT_CJK_ROMAN

HB_AAT_LAYOUT_FEATURE_SELECTOR_DEFAULT_LOWER_CASE

HB_AAT_LAYOUT_FEATURE_SELECTOR_DEFAULT_UPPER_CASE

HB_AAT_LAYOUT_FEATURE_SELECTOR_DESIGN_LEVEL1

HB_AAT_LAYOUT_FEATURE_SELECTOR_DESIGN_LEVEL2

HB_AAT_LAYOUT_FEATURE_SELECTOR_DESIGN_LEVEL3

HB_AAT_LAYOUT_FEATURE_SELECTOR_DESIGN_LEVEL4

HB_AAT_LAYOUT_FEATURE_SELECTOR_DESIGN_LEVEL5

HB_AAT_LAYOUT_FEATURE_SELECTOR_DIAGONAL_FRACTIONS

HB_AAT_LAYOUT_FEATURE_SELECTOR_DIAMOND_ANNOTATION

HB_AAT_LAYOUT_FEATURE_SELECTOR_DINGBATS

HB_AAT_LAYOUT_FEATURE_SELECTOR_DIPHTHONG_LIGATURES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_DIPHTHONG_LIGATURES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_DISPLAY_TEXT

HB_AAT_LAYOUT_FEATURE_SELECTOR_ENGRAVED_TEXT

HB_AAT_LAYOUT_FEATURE_SELECTOR_EXPERT_CHARACTERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_EXPONENTS_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_EXPONENTS_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_FLEURONS

HB_AAT_LAYOUT_FEATURE_SELECTOR_FORM_INTERROBANG_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_FORM_INTERROBANG_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_FULL_WIDTH_CJK_ROMAN

HB_AAT_LAYOUT_FEATURE_SELECTOR_FULL_WIDTH_IDEOGRAPHS

HB_AAT_LAYOUT_FEATURE_SELECTOR_FULL_WIDTH_KANA

HB_AAT_LAYOUT_FEATURE_SELECTOR_HALF_WIDTH_CJK_ROMAN

HB_AAT_LAYOUT_FEATURE_SELECTOR_HALF_WIDTH_IDEOGRAPHS

HB_AAT_LAYOUT_FEATURE_SELECTOR_HALF_WIDTH_TEXT

HB_AAT_LAYOUT_FEATURE_SELECTOR_HANJA_TO_HANGUL

HB_AAT_LAYOUT_FEATURE_SELECTOR_HANJA_TO_HANGUL_ALT_ONE

HB_AAT_LAYOUT_FEATURE_SELECTOR_HANJA_TO_HANGUL_ALT_THREE

HB_AAT_LAYOUT_FEATURE_SELECTOR_HANJA_TO_HANGUL_ALT_TWO

HB_AAT_LAYOUT_FEATURE_SELECTOR_HIDE_DIACRITICS

HB_AAT_LAYOUT_FEATURE_SELECTOR_HIRAGANA_TO_KATAKANA

HB_AAT_LAYOUT_FEATURE_SELECTOR_HISTORICAL_LIGATURES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_HISTORICAL_LIGATURES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_HOJO_CHARACTERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_HYPHENS_TO_EM_DASH_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_HYPHENS_TO_EM_DASH_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_HYPHEN_TO_EN_DASH_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_HYPHEN_TO_EN_DASH_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_HYPHEN_TO_MINUS_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_HYPHEN_TO_MINUS_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_IDEOGRAPHIC_ALT_FIVE

HB_AAT_LAYOUT_FEATURE_SELECTOR_IDEOGRAPHIC_ALT_FOUR

HB_AAT_LAYOUT_FEATURE_SELECTOR_IDEOGRAPHIC_ALT_ONE

HB_AAT_LAYOUT_FEATURE_SELECTOR_IDEOGRAPHIC_ALT_THREE

HB_AAT_LAYOUT_FEATURE_SELECTOR_IDEOGRAPHIC_ALT_TWO

HB_AAT_LAYOUT_FEATURE_SELECTOR_ILLUMINATED_CAPS

HB_AAT_LAYOUT_FEATURE_SELECTOR_INEQUALITY_LIGATURES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_INEQUALITY_LIGATURES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_INFERIORS

HB_AAT_LAYOUT_FEATURE_SELECTOR_INITIAL_CAPS

HB_AAT_LAYOUT_FEATURE_SELECTOR_INITIAL_CAPS_AND_SMALL_CAPS

HB_AAT_LAYOUT_FEATURE_SELECTOR_INTERNATIONAL_SYMBOLS

HB_AAT_LAYOUT_FEATURE_SELECTOR_INVALID

HB_AAT_LAYOUT_FEATURE_SELECTOR_INVERTED_BOX_ANNOTATION

HB_AAT_LAYOUT_FEATURE_SELECTOR_INVERTED_CIRCLE_ANNOTATION

HB_AAT_LAYOUT_FEATURE_SELECTOR_INVERTED_ROUNDED_BOX_ANNOTATION

HB_AAT_LAYOUT_FEATURE_SELECTOR_JIS1978_CHARACTERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_JIS1983_CHARACTERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_JIS1990_CHARACTERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_JIS2004_CHARACTERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_KANA_TO_ROMANIZATION

HB_AAT_LAYOUT_FEATURE_SELECTOR_KATAKANA_TO_HIRAGANA

HB_AAT_LAYOUT_FEATURE_SELECTOR_LINE_FINAL_SWASHES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_LINE_FINAL_SWASHES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_LINE_INITIAL_SWASHES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_LINE_INITIAL_SWASHES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_LINGUISTIC_REARRANGEMENT_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_LINGUISTIC_REARRANGEMENT_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_LOGOS_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_LOGOS_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_LOWER_CASE_NUMBERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_LOWER_CASE_PETITE_CAPS

HB_AAT_LAYOUT_FEATURE_SELECTOR_LOWER_CASE_SMALL_CAPS

HB_AAT_LAYOUT_FEATURE_SELECTOR_MATHEMATICAL_GREEK_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_MATHEMATICAL_GREEK_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_MATH_SYMBOLS

HB_AAT_LAYOUT_FEATURE_SELECTOR_MONOSPACED_NUMBERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_MONOSPACED_TEXT

HB_AAT_LAYOUT_FEATURE_SELECTOR_NLCCHARACTERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_NON_FINAL_SWASHES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_NON_FINAL_SWASHES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_NORMAL_POSITION

HB_AAT_LAYOUT_FEATURE_SELECTOR_NO_ALTERNATES

HB_AAT_LAYOUT_FEATURE_SELECTOR_NO_ANNOTATION

HB_AAT_LAYOUT_FEATURE_SELECTOR_NO_CJK_ITALIC_ROMAN

HB_AAT_LAYOUT_FEATURE_SELECTOR_NO_CJK_SYMBOL_ALTERNATIVES

HB_AAT_LAYOUT_FEATURE_SELECTOR_NO_FRACTIONS

HB_AAT_LAYOUT_FEATURE_SELECTOR_NO_IDEOGRAPHIC_ALTERNATIVES

HB_AAT_LAYOUT_FEATURE_SELECTOR_NO_ORNAMENTS

HB_AAT_LAYOUT_FEATURE_SELECTOR_NO_RUBY_KANA

HB_AAT_LAYOUT_FEATURE_SELECTOR_NO_STYLE_OPTIONS

HB_AAT_LAYOUT_FEATURE_SELECTOR_NO_STYLISTIC_ALTERNATES

HB_AAT_LAYOUT_FEATURE_SELECTOR_NO_TRANSLITERATION

HB_AAT_LAYOUT_FEATURE_SELECTOR_ORDINALS

HB_AAT_LAYOUT_FEATURE_SELECTOR_PARENTHESIS_ANNOTATION

HB_AAT_LAYOUT_FEATURE_SELECTOR_PARTIALLY_CONNECTED

HB_AAT_LAYOUT_FEATURE_SELECTOR_PERIODS_TO_ELLIPSIS_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_PERIODS_TO_ELLIPSIS_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_PERIOD_ANNOTATION

HB_AAT_LAYOUT_FEATURE_SELECTOR_PI_CHARACTERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_PREVENT_OVERLAP_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_PREVENT_OVERLAP_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_PROPORTIONAL_CJK_ROMAN

HB_AAT_LAYOUT_FEATURE_SELECTOR_PROPORTIONAL_IDEOGRAPHS

HB_AAT_LAYOUT_FEATURE_SELECTOR_PROPORTIONAL_KANA

HB_AAT_LAYOUT_FEATURE_SELECTOR_PROPORTIONAL_NUMBERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_PROPORTIONAL_TEXT

HB_AAT_LAYOUT_FEATURE_SELECTOR_QUARTER_WIDTH_NUMBERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_QUARTER_WIDTH_TEXT

HB_AAT_LAYOUT_FEATURE_SELECTOR_RARE_LIGATURES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_RARE_LIGATURES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_REBUS_PICTURES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_REBUS_PICTURES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_REQUIRED_LIGATURES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_REQUIRED_LIGATURES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_ROMANIZATION_TO_HIRAGANA

HB_AAT_LAYOUT_FEATURE_SELECTOR_ROMANIZATION_TO_KATAKANA

HB_AAT_LAYOUT_FEATURE_SELECTOR_ROMAN_NUMERAL_ANNOTATION

HB_AAT_LAYOUT_FEATURE_SELECTOR_ROUNDED_BOX_ANNOTATION

HB_AAT_LAYOUT_FEATURE_SELECTOR_RUBY_KANA

HB_AAT_LAYOUT_FEATURE_SELECTOR_RUBY_KANA_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_RUBY_KANA_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_SCIENTIFIC_INFERIORS

HB_AAT_LAYOUT_FEATURE_SELECTOR_SHOW_DIACRITICS

HB_AAT_LAYOUT_FEATURE_SELECTOR_SIMPLIFIED_CHARACTERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_SLASHED_ZERO_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_SLASHED_ZERO_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_SLASH_TO_DIVIDE_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_SLASH_TO_DIVIDE_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_SMALL_CAPS

HB_AAT_LAYOUT_FEATURE_SELECTOR_SMART_QUOTES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_SMART_QUOTES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_SQUARED_LIGATURES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_SQUARED_LIGATURES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_EIGHTEEN_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_EIGHTEEN_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_EIGHT_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_EIGHT_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_ELEVEN_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_ELEVEN_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_FIFTEEN_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_FIFTEEN_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_FIVE_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_FIVE_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_FOURTEEN_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_FOURTEEN_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_FOUR_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_FOUR_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_NINETEEN_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_NINETEEN_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_NINE_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_NINE_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_ONE_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_ONE_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_SEVENTEEN_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_SEVENTEEN_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_SEVEN_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_SEVEN_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_SIXTEEN_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_SIXTEEN_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_SIX_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_SIX_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_TEN_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_TEN_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_THIRTEEN_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_THIRTEEN_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_THREE_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_THREE_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_TWELVE_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_TWELVE_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_TWENTY_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_TWENTY_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_TWO_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_STYLISTIC_ALT_TWO_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_SUBSTITUTE_VERTICAL_FORMS_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_SUBSTITUTE_VERTICAL_FORMS_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_SUPERIORS

HB_AAT_LAYOUT_FEATURE_SELECTOR_SWASH_ALTERNATES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_SWASH_ALTERNATES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_SYMBOL_LIGATURES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_SYMBOL_LIGATURES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_TALL_CAPS

HB_AAT_LAYOUT_FEATURE_SELECTOR_THIRD_WIDTH_NUMBERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_THIRD_WIDTH_TEXT

HB_AAT_LAYOUT_FEATURE_SELECTOR_TITLING_CAPS

HB_AAT_LAYOUT_FEATURE_SELECTOR_TRADITIONAL_ALT_FIVE

HB_AAT_LAYOUT_FEATURE_SELECTOR_TRADITIONAL_ALT_FOUR

HB_AAT_LAYOUT_FEATURE_SELECTOR_TRADITIONAL_ALT_ONE

HB_AAT_LAYOUT_FEATURE_SELECTOR_TRADITIONAL_ALT_THREE

HB_AAT_LAYOUT_FEATURE_SELECTOR_TRADITIONAL_ALT_TWO

HB_AAT_LAYOUT_FEATURE_SELECTOR_TRADITIONAL_CHARACTERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_TRADITIONAL_NAMES_CHARACTERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_TRANSCODING_COMPOSITION_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_TRANSCODING_COMPOSITION_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_UNCONNECTED

HB_AAT_LAYOUT_FEATURE_SELECTOR_UPPER_AND_LOWER_CASE

HB_AAT_LAYOUT_FEATURE_SELECTOR_UPPER_CASE_NUMBERS

HB_AAT_LAYOUT_FEATURE_SELECTOR_UPPER_CASE_PETITE_CAPS

HB_AAT_LAYOUT_FEATURE_SELECTOR_UPPER_CASE_SMALL_CAPS

HB_AAT_LAYOUT_FEATURE_SELECTOR_VERTICAL_FRACTIONS

HB_AAT_LAYOUT_FEATURE_SELECTOR_WORD_FINAL_SWASHES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_WORD_FINAL_SWASHES_ON

HB_AAT_LAYOUT_FEATURE_SELECTOR_WORD_INITIAL_SWASHES_OFF

HB_AAT_LAYOUT_FEATURE_SELECTOR_WORD_INITIAL_SWASHES_ON

HB_AAT_LAYOUT_FEATURE_TYPE_ALL_TYPOGRAPHIC

HB_AAT_LAYOUT_FEATURE_TYPE_ALTERNATE_KANA

HB_AAT_LAYOUT_FEATURE_TYPE_ANNOTATION_TYPE

HB_AAT_LAYOUT_FEATURE_TYPE_CASE_SENSITIVE_LAYOUT

HB_AAT_LAYOUT_FEATURE_TYPE_CHARACTER_ALTERNATIVES

HB_AAT_LAYOUT_FEATURE_TYPE_CHARACTER_SHAPE

HB_AAT_LAYOUT_FEATURE_TYPE_CJK_ROMAN_SPACING_TYPE

HB_AAT_LAYOUT_FEATURE_TYPE_CJK_SYMBOL_ALTERNATIVES_TYPE

HB_AAT_LAYOUT_FEATURE_TYPE_CJK_VERTICAL_ROMAN_PLACEMENT_TYPE

HB_AAT_LAYOUT_FEATURE_TYPE_CONTEXTUAL_ALTERNATIVES

HB_AAT_LAYOUT_FEATURE_TYPE_CURISVE_CONNECTION

HB_AAT_LAYOUT_FEATURE_TYPE_DESIGN_COMPLEXITY_TYPE

HB_AAT_LAYOUT_FEATURE_TYPE_DIACRITICS_TYPE

HB_AAT_LAYOUT_FEATURE_TYPE_FRACTIONS

HB_AAT_LAYOUT_FEATURE_TYPE_IDEOGRAPHIC_ALTERNATIVES_TYPE

HB_AAT_LAYOUT_FEATURE_TYPE_IDEOGRAPHIC_SPACING_TYPE

HB_AAT_LAYOUT_FEATURE_TYPE_INVALID

HB_AAT_LAYOUT_FEATURE_TYPE_ITALIC_CJK_ROMAN

HB_AAT_LAYOUT_FEATURE_TYPE_KANA_SPACING_TYPE

HB_AAT_LAYOUT_FEATURE_TYPE_LANGUAGE_TAG_TYPE

HB_AAT_LAYOUT_FEATURE_TYPE_LETTER_CASE

HB_AAT_LAYOUT_FEATURE_TYPE_LIGATURES

HB_AAT_LAYOUT_FEATURE_TYPE_LINGUISTIC_REARRANGEMENT

HB_AAT_LAYOUT_FEATURE_TYPE_LOWER_CASE

HB_AAT_LAYOUT_FEATURE_TYPE_MATHEMATICAL_EXTRAS

HB_AAT_LAYOUT_FEATURE_TYPE_NUMBER_CASE

HB_AAT_LAYOUT_FEATURE_TYPE_NUMBER_SPACING

HB_AAT_LAYOUT_FEATURE_TYPE_ORNAMENT_SETS_TYPE

HB_AAT_LAYOUT_FEATURE_TYPE_OVERLAPPING_CHARACTERS_TYPE

HB_AAT_LAYOUT_FEATURE_TYPE_RUBY_KANA

HB_AAT_LAYOUT_FEATURE_TYPE_SMART_SWASH_TYPE

HB_AAT_LAYOUT_FEATURE_TYPE_STYLE_OPTIONS

HB_AAT_LAYOUT_FEATURE_TYPE_STYLISTIC_ALTERNATIVES

HB_AAT_LAYOUT_FEATURE_TYPE_TEXT_SPACING

HB_AAT_LAYOUT_FEATURE_TYPE_TRANSLITERATION

HB_AAT_LAYOUT_FEATURE_TYPE_TYPOGRAPHIC_EXTRAS

HB_AAT_LAYOUT_FEATURE_TYPE_UNICODE_DECOMPOSITION_TYPE

HB_AAT_LAYOUT_FEATURE_TYPE_UPPER_CASE

HB_AAT_LAYOUT_FEATURE_TYPE_VERTICAL_POSITION

HB_AAT_LAYOUT_FEATURE_TYPE_VERTICAL_SUBSTITUTION

HB_AAT_LAYOUT_NO_SELECTOR_INDEX

HB_BUFFER_CLUSTER_LEVEL_CHARACTERS

HB_BUFFER_CLUSTER_LEVEL_DEFAULT

HB_BUFFER_CLUSTER_LEVEL_MONOTONE_CHARACTERS

HB_BUFFER_CLUSTER_LEVEL_MONOTONE_GRAPHEMES

HB_BUFFER_CONTENT_TYPE_GLYPHS

HB_BUFFER_CONTENT_TYPE_INVALID

HB_BUFFER_CONTENT_TYPE_UNICODE

HB_BUFFER_DIFF_FLAG_CLUSTER_MISMATCH

HB_BUFFER_DIFF_FLAG_CODEPOINT_MISMATCH

HB_BUFFER_DIFF_FLAG_CONTENT_TYPE_MISMATCH

HB_BUFFER_DIFF_FLAG_DOTTED_CIRCLE_PRESENT

HB_BUFFER_DIFF_FLAG_EQUAL

HB_BUFFER_DIFF_FLAG_GLYPH_FLAGS_MISMATCH

HB_BUFFER_DIFF_FLAG_LENGTH_MISMATCH

HB_BUFFER_DIFF_FLAG_NOTDEF_PRESENT

HB_BUFFER_DIFF_FLAG_POSITION_MISMATCH

HB_BUFFER_FLAG_BOT

HB_BUFFER_FLAG_DEFAULT

HB_BUFFER_FLAG_DO_NOT_INSERT_DOTTED_CIRCLE

HB_BUFFER_FLAG_EOT

HB_BUFFER_FLAG_PRESERVE_DEFAULT_IGNORABLES

HB_BUFFER_FLAG_REMOVE_DEFAULT_IGNORABLES

HB_BUFFER_REPLACEMENT_CODEPOINT_DEFAULT

HB_BUFFER_SERIALIZE_FLAG_DEFAULT

HB_BUFFER_SERIALIZE_FLAG_GLYPH_EXTENTS

HB_BUFFER_SERIALIZE_FLAG_GLYPH_FLAGS

HB_BUFFER_SERIALIZE_FLAG_NO_ADVANCES

HB_BUFFER_SERIALIZE_FLAG_NO_CLUSTERS

HB_BUFFER_SERIALIZE_FLAG_NO_GLYPH_NAMES

HB_BUFFER_SERIALIZE_FLAG_NO_POSITIONS

HB_BUFFER_SERIALIZE_FORMAT_INVALID

HB_BUFFER_SERIALIZE_FORMAT_JSON

HB_BUFFER_SERIALIZE_FORMAT_TEXT

HB_DIRECTION_BTT

HB_DIRECTION_INVALID

HB_DIRECTION_LTR

HB_DIRECTION_RTL

HB_DIRECTION_TTB

HB_FEATURE_GLOBAL_START

HB_GLYPH_FLAG_DEFINED

HB_GLYPH_FLAG_UNSAFE_TO_BREAK

HB_MEMORY_MODE_DUPLICATE

HB_MEMORY_MODE_READONLY

HB_MEMORY_MODE_READONLY_MAY_MAKE_WRITABLE

HB_MEMORY_MODE_WRITABLE

HB_OT_COLOR_PALETTE_FLAG_DEFAULT

HB_OT_COLOR_PALETTE_FLAG_USABLE_WITH_DARK_BACKGROUND

HB_OT_COLOR_PALETTE_FLAG_USABLE_WITH_LIGHT_BACKGROUND

HB_OT_LAYOUT_BASELINE_TAG_HANGING

HB_OT_LAYOUT_BASELINE_TAG_IDEO_EMBOX_BOTTOM_OR_LEFT

HB_OT_LAYOUT_BASELINE_TAG_IDEO_EMBOX_TOP_OR_RIGHT

HB_OT_LAYOUT_BASELINE_TAG_IDEO_FACE_BOTTOM_OR_LEFT

HB_OT_LAYOUT_BASELINE_TAG_IDEO_FACE_TOP_OR_RIGHT

HB_OT_LAYOUT_BASELINE_TAG_MATH

HB_OT_LAYOUT_BASELINE_TAG_ROMAN

HB_OT_LAYOUT_DEFAULT_LANGUAGE_INDEX

HB_OT_LAYOUT_GLYPH_CLASS_BASE_GLYPH

HB_OT_LAYOUT_GLYPH_CLASS_COMPONENT

HB_OT_LAYOUT_GLYPH_CLASS_LIGATURE

HB_OT_LAYOUT_GLYPH_CLASS_MARK

HB_OT_LAYOUT_GLYPH_CLASS_UNCLASSIFIED

HB_OT_LAYOUT_NO_FEATURE_INDEX

HB_OT_LAYOUT_NO_SCRIPT_INDEX

HB_OT_LAYOUT_NO_VARIATIONS_INDEX

HB_OT_MATH_CONSTANT_ACCENT_BASE_HEIGHT

HB_OT_MATH_CONSTANT_AXIS_HEIGHT

HB_OT_MATH_CONSTANT_DELIMITED_SUB_FORMULA_MIN_HEIGHT

HB_OT_MATH_CONSTANT_DISPLAY_OPERATOR_MIN_HEIGHT

HB_OT_MATH_CONSTANT_FLATTENED_ACCENT_BASE_HEIGHT

HB_OT_MATH_CONSTANT_FRACTION_DENOMINATOR_DISPLAY_STYLE_SHIFT_DOWN

HB_OT_MATH_CONSTANT_FRACTION_DENOMINATOR_GAP_MIN

HB_OT_MATH_CONSTANT_FRACTION_DENOMINATOR_SHIFT_DOWN

HB_OT_MATH_CONSTANT_FRACTION_DENOM_DISPLAY_STYLE_GAP_MIN

HB_OT_MATH_CONSTANT_FRACTION_NUMERATOR_DISPLAY_STYLE_SHIFT_UP

HB_OT_MATH_CONSTANT_FRACTION_NUMERATOR_GAP_MIN

HB_OT_MATH_CONSTANT_FRACTION_NUMERATOR_SHIFT_UP

HB_OT_MATH_CONSTANT_FRACTION_NUM_DISPLAY_STYLE_GAP_MIN

HB_OT_MATH_CONSTANT_FRACTION_RULE_THICKNESS

HB_OT_MATH_CONSTANT_LOWER_LIMIT_BASELINE_DROP_MIN

HB_OT_MATH_CONSTANT_LOWER_LIMIT_GAP_MIN

HB_OT_MATH_CONSTANT_MATH_LEADING

HB_OT_MATH_CONSTANT_OVERBAR_EXTRA_ASCENDER

HB_OT_MATH_CONSTANT_OVERBAR_RULE_THICKNESS

HB_OT_MATH_CONSTANT_OVERBAR_VERTICAL_GAP

HB_OT_MATH_CONSTANT_RADICAL_DEGREE_BOTTOM_RAISE_PERCENT

HB_OT_MATH_CONSTANT_RADICAL_DISPLAY_STYLE_VERTICAL_GAP

HB_OT_MATH_CONSTANT_RADICAL_EXTRA_ASCENDER

HB_OT_MATH_CONSTANT_RADICAL_KERN_AFTER_DEGREE

HB_OT_MATH_CONSTANT_RADICAL_KERN_BEFORE_DEGREE

HB_OT_MATH_CONSTANT_RADICAL_RULE_THICKNESS

HB_OT_MATH_CONSTANT_RADICAL_VERTICAL_GAP

HB_OT_MATH_CONSTANT_SCRIPT_PERCENT_SCALE_DOWN

HB_OT_MATH_CONSTANT_SCRIPT_SCRIPT_PERCENT_SCALE_DOWN

HB_OT_MATH_CONSTANT_SKEWED_FRACTION_HORIZONTAL_GAP

HB_OT_MATH_CONSTANT_SKEWED_FRACTION_VERTICAL_GAP

HB_OT_MATH_CONSTANT_SPACE_AFTER_SCRIPT

HB_OT_MATH_CONSTANT_STACK_BOTTOM_DISPLAY_STYLE_SHIFT_DOWN

HB_OT_MATH_CONSTANT_STACK_BOTTOM_SHIFT_DOWN

HB_OT_MATH_CONSTANT_STACK_DISPLAY_STYLE_GAP_MIN

HB_OT_MATH_CONSTANT_STACK_GAP_MIN

HB_OT_MATH_CONSTANT_STACK_TOP_DISPLAY_STYLE_SHIFT_UP

HB_OT_MATH_CONSTANT_STACK_TOP_SHIFT_UP

HB_OT_MATH_CONSTANT_STRETCH_STACK_BOTTOM_SHIFT_DOWN

HB_OT_MATH_CONSTANT_STRETCH_STACK_GAP_ABOVE_MIN

HB_OT_MATH_CONSTANT_STRETCH_STACK_GAP_BELOW_MIN

HB_OT_MATH_CONSTANT_STRETCH_STACK_TOP_SHIFT_UP

HB_OT_MATH_CONSTANT_SUBSCRIPT_BASELINE_DROP_MIN

HB_OT_MATH_CONSTANT_SUBSCRIPT_SHIFT_DOWN

HB_OT_MATH_CONSTANT_SUBSCRIPT_TOP_MAX

HB_OT_MATH_CONSTANT_SUB_SUPERSCRIPT_GAP_MIN

HB_OT_MATH_CONSTANT_SUPERSCRIPT_BASELINE_DROP_MAX

HB_OT_MATH_CONSTANT_SUPERSCRIPT_BOTTOM_MAX_WITH_SUBSCRIPT

HB_OT_MATH_CONSTANT_SUPERSCRIPT_BOTTOM_MIN

HB_OT_MATH_CONSTANT_SUPERSCRIPT_SHIFT_UP

HB_OT_MATH_CONSTANT_SUPERSCRIPT_SHIFT_UP_CRAMPED

HB_OT_MATH_CONSTANT_UNDERBAR_EXTRA_DESCENDER

HB_OT_MATH_CONSTANT_UNDERBAR_RULE_THICKNESS

HB_OT_MATH_CONSTANT_UNDERBAR_VERTICAL_GAP

HB_OT_MATH_CONSTANT_UPPER_LIMIT_BASELINE_RISE_MIN

HB_OT_MATH_CONSTANT_UPPER_LIMIT_GAP_MIN

HB_OT_MATH_GLYPH_PART_FLAG_EXTENDER

HB_OT_MATH_KERN_BOTTOM_LEFT

HB_OT_MATH_KERN_BOTTOM_RIGHT

HB_OT_MATH_KERN_TOP_LEFT

HB_OT_MATH_KERN_TOP_RIGHT

HB_OT_MAX_TAGS_PER_LANGUAGE

HB_OT_MAX_TAGS_PER_SCRIPT

HB_OT_META_TAG_DESIGN_LANGUAGES

HB_OT_META_TAG_SUPPORTED_LANGUAGES

HB_OT_METRICS_TAG_CAP_HEIGHT

HB_OT_METRICS_TAG_HORIZONTAL_ASCENDER

HB_OT_METRICS_TAG_HORIZONTAL_CARET_OFFSET

HB_OT_METRICS_TAG_HORIZONTAL_CARET_RISE

HB_OT_METRICS_TAG_HORIZONTAL_CARET_RUN

HB_OT_METRICS_TAG_HORIZONTAL_CLIPPING_ASCENT

HB_OT_METRICS_TAG_HORIZONTAL_CLIPPING_DESCENT

HB_OT_METRICS_TAG_HORIZONTAL_DESCENDER

HB_OT_METRICS_TAG_HORIZONTAL_LINE_GAP

HB_OT_METRICS_TAG_STRIKEOUT_OFFSET

HB_OT_METRICS_TAG_STRIKEOUT_SIZE

HB_OT_METRICS_TAG_SUBSCRIPT_EM_X_OFFSET

HB_OT_METRICS_TAG_SUBSCRIPT_EM_X_SIZE

HB_OT_METRICS_TAG_SUBSCRIPT_EM_Y_OFFSET

HB_OT_METRICS_TAG_SUBSCRIPT_EM_Y_SIZE

HB_OT_METRICS_TAG_SUPERSCRIPT_EM_X_OFFSET

HB_OT_METRICS_TAG_SUPERSCRIPT_EM_X_SIZE

HB_OT_METRICS_TAG_SUPERSCRIPT_EM_Y_OFFSET

HB_OT_METRICS_TAG_SUPERSCRIPT_EM_Y_SIZE

HB_OT_METRICS_TAG_UNDERLINE_OFFSET

HB_OT_METRICS_TAG_UNDERLINE_SIZE

HB_OT_METRICS_TAG_VERTICAL_ASCENDER

HB_OT_METRICS_TAG_VERTICAL_CARET_OFFSET

HB_OT_METRICS_TAG_VERTICAL_CARET_RISE

HB_OT_METRICS_TAG_VERTICAL_CARET_RUN

HB_OT_METRICS_TAG_VERTICAL_DESCENDER

HB_OT_METRICS_TAG_VERTICAL_LINE_GAP

HB_OT_METRICS_TAG_X_HEIGHT

HB_OT_NAME_ID_CID_FINDFONT_NAME

HB_OT_NAME_ID_COPYRIGHT

HB_OT_NAME_ID_DARK_BACKGROUND

HB_OT_NAME_ID_DESCRIPTION

HB_OT_NAME_ID_DESIGNER

HB_OT_NAME_ID_DESIGNER_URL

HB_OT_NAME_ID_FONT_FAMILY

HB_OT_NAME_ID_FONT_SUBFAMILY

HB_OT_NAME_ID_FULL_NAME

HB_OT_NAME_ID_INVALID

HB_OT_NAME_ID_LICENSE

HB_OT_NAME_ID_LICENSE_URL

HB_OT_NAME_ID_LIGHT_BACKGROUND

HB_OT_NAME_ID_MAC_FULL_NAME

HB_OT_NAME_ID_MANUFACTURER

HB_OT_NAME_ID_POSTSCRIPT_NAME

HB_OT_NAME_ID_SAMPLE_TEXT

HB_OT_NAME_ID_TRADEMARK

HB_OT_NAME_ID_TYPOGRAPHIC_FAMILY

HB_OT_NAME_ID_TYPOGRAPHIC_SUBFAMILY

HB_OT_NAME_ID_UNIQUE_ID

HB_OT_NAME_ID_VARIATIONS_PS_PREFIX

HB_OT_NAME_ID_VENDOR_URL

HB_OT_NAME_ID_VERSION_STRING

HB_OT_NAME_ID_WWS_FAMILY

HB_OT_NAME_ID_WWS_SUBFAMILY

HB_OT_VAR_AXIS_FLAG_HIDDEN

HB_SCRIPT_ADLAM

HB_SCRIPT_AHOM

HB_SCRIPT_ANATOLIAN_HIEROGLYPHS

HB_SCRIPT_ARABIC

HB_SCRIPT_ARMENIAN

HB_SCRIPT_AVESTAN

HB_SCRIPT_BALINESE

HB_SCRIPT_BAMUM

HB_SCRIPT_BASSA_VAH

HB_SCRIPT_BATAK

HB_SCRIPT_BENGALI

HB_SCRIPT_BHAIKSUKI

HB_SCRIPT_BOPOMOFO

HB_SCRIPT_BRAHMI

HB_SCRIPT_BRAILLE

HB_SCRIPT_BUGINESE

HB_SCRIPT_BUHID

HB_SCRIPT_CANADIAN_SYLLABICS

HB_SCRIPT_CARIAN

HB_SCRIPT_CAUCASIAN_ALBANIAN

HB_SCRIPT_CHAKMA

HB_SCRIPT_CHAM

HB_SCRIPT_CHEROKEE

HB_SCRIPT_COMMON

HB_SCRIPT_COPTIC

HB_SCRIPT_CUNEIFORM

HB_SCRIPT_CYPRIOT

HB_SCRIPT_CYRILLIC

HB_SCRIPT_DESERET

HB_SCRIPT_DEVANAGARI

HB_SCRIPT_DOGRA

HB_SCRIPT_DUPLOYAN

HB_SCRIPT_EGYPTIAN_HIEROGLYPHS

HB_SCRIPT_ELBASAN

HB_SCRIPT_ELYMAIC

HB_SCRIPT_ETHIOPIC

HB_SCRIPT_GEORGIAN

HB_SCRIPT_GLAGOLITIC

HB_SCRIPT_GOTHIC

HB_SCRIPT_GRANTHA

HB_SCRIPT_GREEK

HB_SCRIPT_GUJARATI

HB_SCRIPT_GUNJALA_GONDI

HB_SCRIPT_GURMUKHI

HB_SCRIPT_HAN

HB_SCRIPT_HANGUL

HB_SCRIPT_HANIFI_ROHINGYA

HB_SCRIPT_HANUNOO

HB_SCRIPT_HATRAN

HB_SCRIPT_HEBREW

HB_SCRIPT_HIRAGANA

HB_SCRIPT_IMPERIAL_ARAMAIC

HB_SCRIPT_INHERITED

HB_SCRIPT_INSCRIPTIONAL_PAHLAVI

HB_SCRIPT_INSCRIPTIONAL_PARTHIAN

HB_SCRIPT_INVALID

HB_SCRIPT_JAVANESE

HB_SCRIPT_KAITHI

HB_SCRIPT_KANNADA

HB_SCRIPT_KATAKANA

HB_SCRIPT_KAYAH_LI

HB_SCRIPT_KHAROSHTHI

HB_SCRIPT_KHMER

HB_SCRIPT_KHOJKI

HB_SCRIPT_KHUDAWADI

HB_SCRIPT_LAO

HB_SCRIPT_LATIN

HB_SCRIPT_LEPCHA

HB_SCRIPT_LIMBU

HB_SCRIPT_LINEAR_A

HB_SCRIPT_LINEAR_B

HB_SCRIPT_LISU

HB_SCRIPT_LYCIAN

HB_SCRIPT_LYDIAN

HB_SCRIPT_MAHAJANI

HB_SCRIPT_MAKASAR

HB_SCRIPT_MALAYALAM

HB_SCRIPT_MANDAIC

HB_SCRIPT_MANICHAEAN

HB_SCRIPT_MARCHEN

HB_SCRIPT_MASARAM_GONDI

HB_SCRIPT_MEDEFAIDRIN

HB_SCRIPT_MEETEI_MAYEK

HB_SCRIPT_MENDE_KIKAKUI

HB_SCRIPT_MEROITIC_CURSIVE

HB_SCRIPT_MEROITIC_HIEROGLYPHS

HB_SCRIPT_MIAO

HB_SCRIPT_MODI

HB_SCRIPT_MONGOLIAN

HB_SCRIPT_MRO

HB_SCRIPT_MULTANI

HB_SCRIPT_MYANMAR

HB_SCRIPT_NABATAEAN

HB_SCRIPT_NANDINAGARI

HB_SCRIPT_NEWA

HB_SCRIPT_NEW_TAI_LUE

HB_SCRIPT_NKO

HB_SCRIPT_NUSHU

HB_SCRIPT_NYIAKENG_PUACHUE_HMONG

HB_SCRIPT_OGHAM

HB_SCRIPT_OLD_HUNGARIAN

HB_SCRIPT_OLD_ITALIC

HB_SCRIPT_OLD_NORTH_ARABIAN

HB_SCRIPT_OLD_PERMIC

HB_SCRIPT_OLD_PERSIAN

HB_SCRIPT_OLD_SOGDIAN

HB_SCRIPT_OLD_SOUTH_ARABIAN

HB_SCRIPT_OLD_TURKIC

HB_SCRIPT_OL_CHIKI

HB_SCRIPT_ORIYA

HB_SCRIPT_OSAGE

HB_SCRIPT_OSMANYA

HB_SCRIPT_PAHAWH_HMONG

HB_SCRIPT_PALMYRENE

HB_SCRIPT_PAU_CIN_HAU

HB_SCRIPT_PHAGS_PA

HB_SCRIPT_PHOENICIAN

HB_SCRIPT_PSALTER_PAHLAVI

HB_SCRIPT_REJANG

HB_SCRIPT_RUNIC

HB_SCRIPT_SAMARITAN

HB_SCRIPT_SAURASHTRA

HB_SCRIPT_SHARADA

HB_SCRIPT_SHAVIAN

HB_SCRIPT_SIDDHAM

HB_SCRIPT_SIGNWRITING

HB_SCRIPT_SINHALA

HB_SCRIPT_SOGDIAN

HB_SCRIPT_SORA_SOMPENG

HB_SCRIPT_SOYOMBO

HB_SCRIPT_SUNDANESE

HB_SCRIPT_SYLOTI_NAGRI

HB_SCRIPT_SYRIAC

HB_SCRIPT_TAGALOG

HB_SCRIPT_TAGBANWA

HB_SCRIPT_TAI_LE

HB_SCRIPT_TAI_THAM

HB_SCRIPT_TAI_VIET

HB_SCRIPT_TAKRI

HB_SCRIPT_TAMIL

HB_SCRIPT_TANGUT

HB_SCRIPT_TELUGU

HB_SCRIPT_THAANA

HB_SCRIPT_THAI

HB_SCRIPT_TIBETAN

HB_SCRIPT_TIFINAGH

HB_SCRIPT_TIRHUTA

HB_SCRIPT_UGARITIC

HB_SCRIPT_UNKNOWN

HB_SCRIPT_VAI

HB_SCRIPT_WANCHO

HB_SCRIPT_WARANG_CITI

HB_SCRIPT_YI

HB_SCRIPT_ZANABAZAR_SQUARE

HB_UNICODE_COMBINING_CLASS_ABOVE

HB_UNICODE_COMBINING_CLASS_ABOVE_LEFT

HB_UNICODE_COMBINING_CLASS_ABOVE_RIGHT

HB_UNICODE_COMBINING_CLASS_ATTACHED_ABOVE

HB_UNICODE_COMBINING_CLASS_ATTACHED_ABOVE_RIGHT

HB_UNICODE_COMBINING_CLASS_ATTACHED_BELOW

HB_UNICODE_COMBINING_CLASS_ATTACHED_BELOW_LEFT

HB_UNICODE_COMBINING_CLASS_BELOW

HB_UNICODE_COMBINING_CLASS_BELOW_LEFT

HB_UNICODE_COMBINING_CLASS_BELOW_RIGHT

HB_UNICODE_COMBINING_CLASS_CCC10

HB_UNICODE_COMBINING_CLASS_CCC11

HB_UNICODE_COMBINING_CLASS_CCC12

HB_UNICODE_COMBINING_CLASS_CCC13

HB_UNICODE_COMBINING_CLASS_CCC14

HB_UNICODE_COMBINING_CLASS_CCC15

HB_UNICODE_COMBINING_CLASS_CCC16

HB_UNICODE_COMBINING_CLASS_CCC17

HB_UNICODE_COMBINING_CLASS_CCC18

HB_UNICODE_COMBINING_CLASS_CCC19

HB_UNICODE_COMBINING_CLASS_CCC20

HB_UNICODE_COMBINING_CLASS_CCC21

HB_UNICODE_COMBINING_CLASS_CCC22

HB_UNICODE_COMBINING_CLASS_CCC23

HB_UNICODE_COMBINING_CLASS_CCC24

HB_UNICODE_COMBINING_CLASS_CCC25

HB_UNICODE_COMBINING_CLASS_CCC26

HB_UNICODE_COMBINING_CLASS_CCC27

HB_UNICODE_COMBINING_CLASS_CCC28

HB_UNICODE_COMBINING_CLASS_CCC29

HB_UNICODE_COMBINING_CLASS_CCC30

HB_UNICODE_COMBINING_CLASS_CCC31

HB_UNICODE_COMBINING_CLASS_CCC32

HB_UNICODE_COMBINING_CLASS_CCC33

HB_UNICODE_COMBINING_CLASS_CCC34

HB_UNICODE_COMBINING_CLASS_CCC35

HB_UNICODE_COMBINING_CLASS_CCC36

HB_UNICODE_COMBINING_CLASS_CCC84

HB_UNICODE_COMBINING_CLASS_CCC91

HB_UNICODE_COMBINING_CLASS_CCC103

HB_UNICODE_COMBINING_CLASS_CCC107

HB_UNICODE_COMBINING_CLASS_CCC118

HB_UNICODE_COMBINING_CLASS_CCC122

HB_UNICODE_COMBINING_CLASS_CCC129

HB_UNICODE_COMBINING_CLASS_CCC130

HB_UNICODE_COMBINING_CLASS_CCC133

HB_UNICODE_COMBINING_CLASS_DOUBLE_ABOVE

HB_UNICODE_COMBINING_CLASS_DOUBLE_BELOW

HB_UNICODE_COMBINING_CLASS_INVALID

HB_UNICODE_COMBINING_CLASS_IOTA_SUBSCRIPT

HB_UNICODE_COMBINING_CLASS_KANA_VOICING

HB_UNICODE_COMBINING_CLASS_LEFT

HB_UNICODE_COMBINING_CLASS_NOT_REORDERED

HB_UNICODE_COMBINING_CLASS_NUKTA

HB_UNICODE_COMBINING_CLASS_OVERLAY

HB_UNICODE_COMBINING_CLASS_RIGHT

HB_UNICODE_COMBINING_CLASS_VIRAMA

HB_UNICODE_GENERAL_CATEGORY_CLOSE_PUNCTUATION

HB_UNICODE_GENERAL_CATEGORY_CONNECT_PUNCTUATION

HB_UNICODE_GENERAL_CATEGORY_CONTROL

HB_UNICODE_GENERAL_CATEGORY_CURRENCY_SYMBOL

HB_UNICODE_GENERAL_CATEGORY_DASH_PUNCTUATION

HB_UNICODE_GENERAL_CATEGORY_DECIMAL_NUMBER

HB_UNICODE_GENERAL_CATEGORY_ENCLOSING_MARK

HB_UNICODE_GENERAL_CATEGORY_FINAL_PUNCTUATION

HB_UNICODE_GENERAL_CATEGORY_FORMAT

HB_UNICODE_GENERAL_CATEGORY_INITIAL_PUNCTUATION

HB_UNICODE_GENERAL_CATEGORY_LETTER_NUMBER

HB_UNICODE_GENERAL_CATEGORY_LINE_SEPARATOR

HB_UNICODE_GENERAL_CATEGORY_LOWERCASE_LETTER

HB_UNICODE_GENERAL_CATEGORY_MATH_SYMBOL

HB_UNICODE_GENERAL_CATEGORY_MODIFIER_LETTER

HB_UNICODE_GENERAL_CATEGORY_MODIFIER_SYMBOL

HB_UNICODE_GENERAL_CATEGORY_NON_SPACING_MARK

HB_UNICODE_GENERAL_CATEGORY_OPEN_PUNCTUATION

HB_UNICODE_GENERAL_CATEGORY_OTHER_LETTER

HB_UNICODE_GENERAL_CATEGORY_OTHER_NUMBER

HB_UNICODE_GENERAL_CATEGORY_OTHER_PUNCTUATION

HB_UNICODE_GENERAL_CATEGORY_OTHER_SYMBOL

HB_UNICODE_GENERAL_CATEGORY_PARAGRAPH_SEPARATOR

HB_UNICODE_GENERAL_CATEGORY_PRIVATE_USE

HB_UNICODE_GENERAL_CATEGORY_SPACE_SEPARATOR

HB_UNICODE_GENERAL_CATEGORY_SPACING_MARK

HB_UNICODE_GENERAL_CATEGORY_SURROGATE

HB_UNICODE_GENERAL_CATEGORY_TITLECASE_LETTER

HB_UNICODE_GENERAL_CATEGORY_UNASSIGNED

HB_UNICODE_GENERAL_CATEGORY_UPPERCASE_LETTER

HB_UNICODE_MAX

_HB_AAT_LAYOUT_FEATURE_SELECTOR_MAX_VALUE

_HB_AAT_LAYOUT_FEATURE_TYPE_MAX_VALUE

_HB_OT_LAYOUT_BASELINE_TAG_MAX_VALUE

_HB_OT_META_TAG_MAX_VALUE

_HB_OT_METRICS_TAG_MAX_VALUE

_HB_OT_VAR_AXIS_FLAG_MAX_VALUE

_HB_SCRIPT_MAX_VALUE

_HB_SCRIPT_MAX_VALUE_SIGNED

Functions

hb_aat_layout_feature_type_get_name_id⚠

hb_aat_layout_feature_type_get_selector_infos⚠

hb_aat_layout_get_feature_types⚠

hb_aat_layout_has_positioning⚠

hb_aat_layout_has_substitution⚠

hb_aat_layout_has_tracking⚠

hb_blob_copy_writable_or_fail⚠

hb_blob_create⚠

hb_blob_create_from_file⚠

hb_blob_create_sub_blob⚠

hb_blob_destroy⚠

hb_blob_get_data⚠

hb_blob_get_data_writable⚠

hb_blob_get_empty⚠

hb_blob_get_length⚠

hb_blob_get_user_data⚠

hb_blob_is_immutable⚠

hb_blob_make_immutable⚠

hb_blob_reference⚠

hb_blob_set_user_data⚠

hb_buffer_add⚠

hb_buffer_add_codepoints⚠

hb_buffer_add_latin1⚠

hb_buffer_add_utf8⚠

hb_buffer_add_utf16⚠

hb_buffer_add_utf32⚠

hb_buffer_allocation_successful⚠

hb_buffer_append⚠

hb_buffer_clear_contents⚠

hb_buffer_create⚠

hb_buffer_deserialize_glyphs⚠

hb_buffer_destroy⚠

hb_buffer_diff⚠

hb_buffer_get_cluster_level⚠

hb_buffer_get_content_type⚠

hb_buffer_get_direction⚠

hb_buffer_get_empty⚠

hb_buffer_get_flags⚠

hb_buffer_get_glyph_infos⚠

hb_buffer_get_glyph_positions⚠

hb_buffer_get_invisible_glyph⚠

hb_buffer_get_language⚠

hb_buffer_get_length⚠

hb_buffer_get_replacement_codepoint⚠

hb_buffer_get_script⚠

hb_buffer_get_segment_properties⚠

hb_buffer_get_unicode_funcs⚠

hb_buffer_get_user_data⚠

hb_buffer_guess_segment_properties⚠

hb_buffer_normalize_glyphs⚠

hb_buffer_pre_allocate⚠

hb_buffer_reference⚠

hb_buffer_reset⚠

hb_buffer_reverse⚠

hb_buffer_reverse_clusters⚠

hb_buffer_reverse_range⚠

hb_buffer_serialize_format_from_string⚠

hb_buffer_serialize_format_to_string⚠

hb_buffer_serialize_glyphs⚠

hb_buffer_serialize_list_formats⚠

hb_buffer_set_cluster_level⚠

hb_buffer_set_content_type⚠

hb_buffer_set_direction⚠

hb_buffer_set_flags⚠

hb_buffer_set_invisible_glyph⚠

hb_buffer_set_language⚠

hb_buffer_set_length⚠

hb_buffer_set_message_func⚠

hb_buffer_set_replacement_codepoint⚠

hb_buffer_set_script⚠

hb_buffer_set_segment_properties⚠

hb_buffer_set_unicode_funcs⚠

hb_buffer_set_user_data⚠

hb_color_get_alpha⚠

hb_color_get_blue⚠

hb_color_get_green⚠

hb_color_get_red⚠

hb_direction_from_string⚠

hb_direction_to_string⚠

hb_face_builder_add_table⚠

hb_face_builder_create⚠

hb_face_collect_unicodes⚠

hb_face_collect_variation_selectors⚠

hb_face_collect_variation_unicodes⚠

hb_face_count⚠

hb_face_create⚠

hb_face_create_for_tables⚠

hb_face_destroy⚠

hb_face_get_empty⚠

hb_face_get_glyph_count⚠

hb_face_get_index⚠

hb_face_get_table_tags⚠

hb_face_get_upem⚠

hb_face_get_user_data⚠

hb_face_is_immutable⚠

hb_face_make_immutable⚠

hb_face_reference⚠

hb_face_reference_blob⚠

hb_face_reference_table⚠

hb_face_set_glyph_count⚠

hb_face_set_index⚠

hb_face_set_upem⚠

hb_face_set_user_data⚠

hb_feature_from_string⚠

hb_feature_to_string⚠

hb_font_add_glyph_origin_for_direction⚠

hb_font_create⚠

hb_font_create_sub_font⚠

hb_font_destroy⚠

hb_font_funcs_create⚠

hb_font_funcs_destroy⚠

hb_font_funcs_get_empty⚠

hb_font_funcs_get_user_data⚠

hb_font_funcs_is_immutable⚠

hb_font_funcs_make_immutable⚠

hb_font_funcs_reference⚠

hb_font_funcs_set_font_h_extents_func⚠
hb_font_funcs_set_font_h_extents_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_font_v_extents_func⚠
hb_font_funcs_set_font_v_extents_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_contour_point_func⚠
hb_font_funcs_set_glyph_contour_point_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_extents_func⚠
hb_font_funcs_set_glyph_extents_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_from_name_func⚠
hb_font_funcs_set_glyph_from_name_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_h_advance_func⚠
hb_font_funcs_set_glyph_h_advance_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_h_advances_func⚠
hb_font_funcs_set_glyph_h_advances_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_h_kerning_func⚠
hb_font_funcs_set_glyph_h_kerning_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_h_origin_func⚠
hb_font_funcs_set_glyph_h_origin_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_name_func⚠
hb_font_funcs_set_glyph_name_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_v_advance_func⚠
hb_font_funcs_set_glyph_v_advance_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_v_advances_func⚠
hb_font_funcs_set_glyph_v_advances_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_glyph_v_origin_func⚠
hb_font_funcs_set_glyph_v_origin_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_nominal_glyph_func⚠
hb_font_funcs_set_nominal_glyph_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_nominal_glyphs_func⚠
hb_font_funcs_set_nominal_glyphs_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_funcs_set_user_data⚠

hb_font_funcs_set_variation_glyph_func⚠
hb_font_funcs_set_variation_glyph_func: @ffuncs: font functions. @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_font_get_empty⚠

hb_font_get_extents_for_direction⚠

hb_font_get_face⚠

hb_font_get_glyph⚠

hb_font_get_glyph_advance_for_direction⚠

hb_font_get_glyph_advances_for_direction⚠

hb_font_get_glyph_contour_point⚠

hb_font_get_glyph_contour_point_for_origin⚠

hb_font_get_glyph_extents⚠

hb_font_get_glyph_extents_for_origin⚠

hb_font_get_glyph_from_name⚠

hb_font_get_glyph_h_advance⚠

hb_font_get_glyph_h_advances⚠

hb_font_get_glyph_h_kerning⚠

hb_font_get_glyph_h_origin⚠

hb_font_get_glyph_kerning_for_direction⚠

hb_font_get_glyph_name⚠

hb_font_get_glyph_origin_for_direction⚠

hb_font_get_glyph_v_advance⚠

hb_font_get_glyph_v_advances⚠

hb_font_get_glyph_v_origin⚠

hb_font_get_h_extents⚠

hb_font_get_nominal_glyph⚠

hb_font_get_parent⚠

hb_font_get_ppem⚠

hb_font_get_ptem⚠

hb_font_get_scale⚠

hb_font_get_user_data⚠

hb_font_get_v_extents⚠

hb_font_get_var_coords_normalized⚠

hb_font_get_variation_glyph⚠

hb_font_glyph_from_string⚠

hb_font_glyph_to_string⚠

hb_font_is_immutable⚠

hb_font_make_immutable⚠

hb_font_reference⚠

hb_font_set_face⚠

hb_font_set_funcs⚠

hb_font_set_funcs_data⚠

hb_font_set_parent⚠

hb_font_set_ppem⚠

hb_font_set_ptem⚠

hb_font_set_scale⚠

hb_font_set_user_data⚠

hb_font_set_var_coords_design⚠

hb_font_set_var_coords_normalized⚠

hb_font_set_variations⚠

hb_font_subtract_glyph_origin_for_direction⚠

hb_ft_font_create_referenced⚠

hb_glyph_info_get_glyph_flags⚠

hb_language_from_string⚠

hb_language_get_default⚠

hb_language_to_string⚠

hb_map_allocation_successful⚠

hb_map_clear⚠

hb_map_create⚠

hb_map_del⚠

hb_map_destroy⚠

hb_map_get⚠

hb_map_get_empty⚠

hb_map_get_population⚠

hb_map_get_user_data⚠

hb_map_has⚠

hb_map_is_empty⚠

hb_map_reference⚠

hb_map_set⚠

hb_map_set_user_data⚠

hb_ot_color_glyph_get_layers⚠

hb_ot_color_glyph_reference_png⚠

hb_ot_color_glyph_reference_svg⚠

hb_ot_color_has_layers⚠

hb_ot_color_has_palettes⚠

hb_ot_color_has_png⚠

hb_ot_color_has_svg⚠

hb_ot_color_palette_color_get_name_id⚠

hb_ot_color_palette_get_colors⚠

hb_ot_color_palette_get_count⚠

hb_ot_color_palette_get_flags⚠

hb_ot_color_palette_get_name_id⚠

hb_ot_font_set_funcs⚠

hb_ot_layout_collect_features⚠

hb_ot_layout_collect_lookups⚠

hb_ot_layout_feature_get_characters⚠

hb_ot_layout_feature_get_lookups⚠

hb_ot_layout_feature_get_name_ids⚠

hb_ot_layout_feature_with_variations_get_lookups⚠

hb_ot_layout_get_attach_points⚠

hb_ot_layout_get_baseline⚠

hb_ot_layout_get_glyph_class⚠

hb_ot_layout_get_glyphs_in_class⚠

hb_ot_layout_get_ligature_carets⚠

hb_ot_layout_get_size_params⚠

hb_ot_layout_has_glyph_classes⚠

hb_ot_layout_has_positioning⚠

hb_ot_layout_has_substitution⚠

hb_ot_layout_language_find_feature⚠

hb_ot_layout_language_get_feature_indexes⚠

hb_ot_layout_language_get_feature_tags⚠

hb_ot_layout_language_get_required_feature⚠

hb_ot_layout_language_get_required_feature_index⚠

hb_ot_layout_lookup_collect_glyphs⚠

hb_ot_layout_lookup_substitute_closure⚠

hb_ot_layout_lookup_would_substitute⚠

hb_ot_layout_lookups_substitute_closure⚠

hb_ot_layout_script_get_language_tags⚠

hb_ot_layout_script_select_language⚠

hb_ot_layout_table_find_feature_variations⚠

hb_ot_layout_table_find_script⚠

hb_ot_layout_table_get_feature_tags⚠

hb_ot_layout_table_get_lookup_count⚠

hb_ot_layout_table_get_script_tags⚠

hb_ot_layout_table_select_script⚠

hb_ot_math_get_constant⚠

hb_ot_math_get_glyph_assembly⚠

hb_ot_math_get_glyph_italics_correction⚠

hb_ot_math_get_glyph_kerning⚠

hb_ot_math_get_glyph_top_accent_attachment⚠

hb_ot_math_get_glyph_variants⚠

hb_ot_math_get_min_connector_overlap⚠

hb_ot_math_has_data⚠

hb_ot_math_is_glyph_extended_shape⚠

hb_ot_meta_get_entry_tags⚠

hb_ot_meta_reference_entry⚠

hb_ot_metrics_get_position⚠

hb_ot_metrics_get_variation⚠

hb_ot_metrics_get_x_variation⚠

hb_ot_metrics_get_y_variation⚠

hb_ot_name_get_utf8⚠

hb_ot_name_get_utf16⚠

hb_ot_name_get_utf32⚠

hb_ot_name_list_names⚠

hb_ot_shape_glyphs_closure⚠

hb_ot_shape_plan_collect_lookups⚠

hb_ot_tag_to_language⚠

hb_ot_tag_to_script⚠

hb_ot_tags_from_script_and_language⚠

hb_ot_tags_to_script_and_language⚠

hb_ot_var_find_axis_info⚠

hb_ot_var_get_axis_count⚠

hb_ot_var_get_axis_infos⚠

hb_ot_var_get_named_instance_count⚠

hb_ot_var_has_data⚠

hb_ot_var_named_instance_get_design_coords⚠

hb_ot_var_named_instance_get_postscript_name_id⚠

hb_ot_var_named_instance_get_subfamily_name_id⚠

hb_ot_var_normalize_coords⚠

hb_ot_var_normalize_variations⚠

hb_script_from_iso15924_tag⚠

hb_script_from_string⚠

hb_script_get_horizontal_direction⚠

hb_script_to_iso15924_tag⚠

hb_segment_properties_equal⚠

hb_segment_properties_hash⚠

hb_set_add⚠

hb_set_add_range⚠

hb_set_allocation_successful⚠

hb_set_clear⚠

hb_set_create⚠

hb_set_del⚠

hb_set_del_range⚠

hb_set_destroy⚠

hb_set_get_empty⚠

hb_set_get_max⚠

hb_set_get_min⚠

hb_set_get_population⚠

hb_set_get_user_data⚠

hb_set_has⚠

hb_set_intersect⚠

hb_set_is_empty⚠

hb_set_is_equal⚠

hb_set_is_subset⚠

hb_set_next⚠

hb_set_next_range⚠

hb_set_previous⚠

hb_set_previous_range⚠

hb_set_reference⚠

hb_set_set⚠

hb_set_set_user_data⚠

hb_set_subtract⚠

hb_set_symmetric_difference⚠

hb_set_union⚠

hb_shape⚠

hb_shape_full⚠

hb_shape_list_shapers⚠

hb_shape_plan_create⚠

hb_shape_plan_create2⚠

hb_shape_plan_create_cached⚠

hb_shape_plan_create_cached2⚠

hb_shape_plan_destroy⚠

hb_shape_plan_execute⚠

hb_shape_plan_get_empty⚠

hb_shape_plan_get_shaper⚠

hb_shape_plan_get_user_data⚠

hb_shape_plan_reference⚠

hb_shape_plan_set_user_data⚠

hb_tag_from_string⚠

hb_tag_to_string⚠

hb_unicode_combining_class⚠
hb_unicode_combining_class:

hb_unicode_compose⚠

hb_unicode_decompose⚠

hb_unicode_funcs_create⚠

hb_unicode_funcs_destroy⚠

hb_unicode_funcs_get_default⚠

hb_unicode_funcs_get_empty⚠

hb_unicode_funcs_get_parent⚠

hb_unicode_funcs_get_user_data⚠

hb_unicode_funcs_is_immutable⚠

hb_unicode_funcs_make_immutable⚠

hb_unicode_funcs_reference⚠

hb_unicode_funcs_set_combining_class_func⚠
hb_unicode_funcs_set_combining_class_func: @ufuncs: a Unicode function structure @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_unicode_funcs_set_compose_func⚠
hb_unicode_funcs_set_compose_func: @ufuncs: a Unicode function structure @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_unicode_funcs_set_decompose_func⚠
hb_unicode_funcs_set_decompose_func: @ufuncs: a Unicode function structure @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_unicode_funcs_set_general_category_func⚠
hb_unicode_funcs_set_general_category_func: @ufuncs: a Unicode function structure @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_unicode_funcs_set_mirroring_func⚠
hb_unicode_funcs_set_mirroring_func: @ufuncs: a Unicode function structure @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_unicode_funcs_set_script_func⚠
hb_unicode_funcs_set_script_func: @ufuncs: a Unicode function structure @func: (closure user_data) (destroy destroy) (scope notified): @user_data: @destroy:

hb_unicode_funcs_set_user_data⚠

hb_unicode_general_category⚠
hb_unicode_general_category:

hb_unicode_mirroring⚠
hb_unicode_mirroring:

hb_unicode_script⚠
hb_unicode_script:

hb_variation_from_string⚠

hb_variation_to_string⚠

hb_version⚠

hb_version_atleast⚠

hb_version_string⚠

Type Definitions

hb_aat_layout_feature_selector_t
hb_aat_layout_feature_selector_t:

hb_aat_layout_feature_type_t
hb_aat_layout_feature_type_t:

hb_bool_t

hb_buffer_cluster_level_t
hb_buffer_cluster_level_t: @HB_BUFFER_CLUSTER_LEVEL_MONOTONE_GRAPHEMES: Return cluster values grouped by graphemes into monotone order. @HB_BUFFER_CLUSTER_LEVEL_MONOTONE_CHARACTERS: Return cluster values grouped into monotone order. @HB_BUFFER_CLUSTER_LEVEL_CHARACTERS: Don't group cluster values. @HB_BUFFER_CLUSTER_LEVEL_DEFAULT: Default cluster level, equal to @HB_BUFFER_CLUSTER_LEVEL_MONOTONE_GRAPHEMES.

hb_buffer_content_type_t
hb_buffer_content_type_t: @HB_BUFFER_CONTENT_TYPE_INVALID: Initial value for new buffer. @HB_BUFFER_CONTENT_TYPE_UNICODE: The buffer contains input characters (before shaping). @HB_BUFFER_CONTENT_TYPE_GLYPHS: The buffer contains output glyphs (after shaping).

hb_buffer_diff_flags_t

hb_buffer_flags_t
hb_buffer_flags_t: @HB_BUFFER_FLAG_DEFAULT: the default buffer flag. @HB_BUFFER_FLAG_BOT: flag indicating that special handling of the beginning of text paragraph can be applied to this buffer. Should usually be set, unless you are passing to the buffer only part of the text without the full context. @HB_BUFFER_FLAG_EOT: flag indicating that special handling of the end of text paragraph can be applied to this buffer, similar to @HB_BUFFER_FLAG_BOT. @HB_BUFFER_FLAG_PRESERVE_DEFAULT_IGNORABLES: flag indication that character with Default_Ignorable Unicode property should use the corresponding glyph from the font, instead of hiding them (done by replacing them with the space glyph and zeroing the advance width.) This flag takes precedence over @HB_BUFFER_FLAG_REMOVE_DEFAULT_IGNORABLES. @HB_BUFFER_FLAG_REMOVE_DEFAULT_IGNORABLES: flag indication that character with Default_Ignorable Unicode property should be removed from glyph string instead of hiding them (done by replacing them with the space glyph and zeroing the advance width.) @HB_BUFFER_FLAG_PRESERVE_DEFAULT_IGNORABLES takes precedence over this flag. Since: 1.8.0 @HB_BUFFER_FLAG_DO_NOT_INSERT_DOTTED_CIRCLE: flag indicating that a dotted circle should not be inserted in the rendering of incorrect character sequences (such at <0905 093E>). Since: 2.4

hb_buffer_message_func_t

hb_buffer_serialize_flags_t
hb_buffer_serialize_flags_t: @HB_BUFFER_SERIALIZE_FLAG_DEFAULT: serialize glyph names, clusters and positions. @HB_BUFFER_SERIALIZE_FLAG_NO_CLUSTERS: do not serialize glyph cluster. @HB_BUFFER_SERIALIZE_FLAG_NO_POSITIONS: do not serialize glyph position information. @HB_BUFFER_SERIALIZE_FLAG_NO_GLYPH_NAMES: do no serialize glyph name. @HB_BUFFER_SERIALIZE_FLAG_GLYPH_EXTENTS: serialize glyph extents. @HB_BUFFER_SERIALIZE_FLAG_GLYPH_FLAGS: serialize glyph flags. Since: 1.5.0 @HB_BUFFER_SERIALIZE_FLAG_NO_ADVANCES: do not serialize glyph advances, glyph offsets will reflect absolute glyph positions. Since: 1.8.0

hb_buffer_serialize_format_t
hb_buffer_serialize_format_t: @HB_BUFFER_SERIALIZE_FORMAT_TEXT: a human-readable, plain text format. @HB_BUFFER_SERIALIZE_FORMAT_JSON: a machine-readable JSON format. @HB_BUFFER_SERIALIZE_FORMAT_INVALID: invalid format.

hb_codepoint_t

hb_color_t
hb_color_t:

hb_destroy_func_t

hb_direction_t
hb_direction_t: @HB_DIRECTION_INVALID: Initial, unset direction. @HB_DIRECTION_LTR: Text is set horizontally from left to right. @HB_DIRECTION_RTL: Text is set horizontally from right to left. @HB_DIRECTION_TTB: Text is set vertically from top to bottom. @HB_DIRECTION_BTT: Text is set vertically from bottom to top.

hb_font_get_font_extents_func_t

hb_font_get_font_h_extents_func_t

hb_font_get_font_v_extents_func_t

hb_font_get_glyph_advance_func_t

hb_font_get_glyph_advances_func_t

hb_font_get_glyph_contour_point_func_t

hb_font_get_glyph_extents_func_t

hb_font_get_glyph_from_name_func_t

hb_font_get_glyph_h_advance_func_t

hb_font_get_glyph_h_advances_func_t

hb_font_get_glyph_h_kerning_func_t

hb_font_get_glyph_h_origin_func_t

hb_font_get_glyph_kerning_func_t

hb_font_get_glyph_name_func_t

hb_font_get_glyph_origin_func_t

hb_font_get_glyph_v_advance_func_t

hb_font_get_glyph_v_advances_func_t

hb_font_get_glyph_v_origin_func_t

hb_font_get_nominal_glyph_func_t

hb_font_get_nominal_glyphs_func_t

hb_font_get_variation_glyph_func_t

hb_glyph_flags_t
hb_glyph_flags_t: @HB_GLYPH_FLAG_UNSAFE_TO_BREAK: Indicates that if input text is broken at the beginning of the cluster this glyph is part of, then both sides need to be re-shaped, as the result might be different. On the flip side, it means that when this flag is not present, then it's safe to break the glyph-run at the beginning of this cluster, and the two sides represent the exact same result one would get if breaking input text at the beginning of this cluster and shaping the two sides separately. This can be used to optimize paragraph layout, by avoiding re-shaping of each line after line-breaking, or limiting the reshaping to a small piece around the breaking point only. @HB_GLYPH_FLAG_DEFINED: All the currently defined flags.

hb_language_t

hb_mask_t

hb_memory_mode_t

hb_ot_color_palette_flags_t
hb_ot_color_palette_flags_t: @HB_OT_COLOR_PALETTE_FLAG_DEFAULT: default indicating that there is nothing special to note about a color palette. @HB_OT_COLOR_PALETTE_FLAG_USABLE_WITH_LIGHT_BACKGROUND: flag indicating that the color palette is appropriate to use when displaying the font on a light background such as white. @HB_OT_COLOR_PALETTE_FLAG_USABLE_WITH_DARK_BACKGROUND: flag indicating that the color palette is appropriate to use when displaying the font on a dark background such as black.

hb_ot_layout_baseline_tag_t
hb_ot_layout_baseline_tag_t: @HB_OT_LAYOUT_BASELINE_TAG_ROMAN: The baseline used by alphabetic scripts such as Latin, Cyrillic and Greek. In vertical writing mode, the alphabetic baseline for characters rotated 90 degrees clockwise. (This would not apply to alphabetic characters that remain upright in vertical writing mode, since these characters are not rotated.) @HB_OT_LAYOUT_BASELINE_TAG_HANGING: The hanging baseline. In horizontal direction, this is the horizontal line from which syllables seem, to hang in Tibetan and other similar scripts. In vertical writing mode, for Tibetan (or some other similar script) characters rotated 90 degrees clockwise. @HB_OT_LAYOUT_BASELINE_TAG_IDEO_FACE_BOTTOM_OR_LEFT: Ideographic character face bottom or left edge, if the direction is horizontal or vertical, respectively. @HB_OT_LAYOUT_BASELINE_TAG_IDEO_FACE_TOP_OR_RIGHT: Ideographic character face top or right edge, if the direction is horizontal or vertical, respectively. @HB_OT_LAYOUT_BASELINE_TAG_IDEO_EMBOX_BOTTOM_OR_LEFT: Ideographic em-box bottom or left edge, if the direction is horizontal or vertical, respectively. @HB_OT_LAYOUT_BASELINE_TAG_IDEO_EMBOX_TOP_OR_RIGHT: Ideographic em-box top or right edge baseline, if the direction is horizontal or vertical, respectively. @HB_OT_LAYOUT_BASELINE_TAG_MATH: The baseline about which mathematical characters are centered. In vertical writing mode when mathematical characters rotated 90 degrees clockwise, are centered.

hb_ot_layout_glyph_class_t
hb_ot_layout_glyph_class_t: @HB_OT_LAYOUT_GLYPH_CLASS_UNCLASSIFIED: Glyphs not matching the other classifications @HB_OT_LAYOUT_GLYPH_CLASS_BASE_GLYPH: Spacing, single characters, capable of accepting marks @HB_OT_LAYOUT_GLYPH_CLASS_LIGATURE: Glyphs that represent ligation of multiple characters @HB_OT_LAYOUT_GLYPH_CLASS_MARK: Non-spacing, combining glyphs that represent marks @HB_OT_LAYOUT_GLYPH_CLASS_COMPONENT: Spacing glyphs that represent part of a single character

hb_ot_math_constant_t
hb_ot_math_constant_t:

hb_ot_math_glyph_part_flags_t
hb_ot_math_glyph_part_flags_t:

hb_ot_math_kern_t
hb_ot_math_kern_t:

hb_ot_meta_tag_t
hb_ot_meta_tag_t: @HB_OT_META_TAG_DESIGN_LANGUAGES: Design languages. Text, using only Basic Latin (ASCII) characters. Indicates languages and/or scripts for the user audiences that the font was primarily designed for. @HB_OT_META_TAG_SUPPORTED_LANGUAGES: Supported languages. Text, using only Basic Latin (ASCII) characters. Indicates languages and/or scripts that the font is declared to be capable of supporting.

hb_ot_metrics_tag_t
hb_ot_metrics_tag_t: @HB_OT_METRICS_TAG_HORIZONTAL_ASCENDER: horizontal ascender. @HB_OT_METRICS_TAG_HORIZONTAL_DESCENDER: horizontal descender. @HB_OT_METRICS_TAG_HORIZONTAL_LINE_GAP: horizontal line gap. @HB_OT_METRICS_TAG_HORIZONTAL_CLIPPING_ASCENT: horizontal clipping ascent. @HB_OT_METRICS_TAG_HORIZONTAL_CLIPPING_DESCENT: horizontal clipping descent. @HB_OT_METRICS_TAG_VERTICAL_ASCENDER: vertical ascender. @HB_OT_METRICS_TAG_VERTICAL_DESCENDER: vertical descender. @HB_OT_METRICS_TAG_VERTICAL_LINE_GAP: vertical line gap. @HB_OT_METRICS_TAG_HORIZONTAL_CARET_RISE: horizontal caret rise. @HB_OT_METRICS_TAG_HORIZONTAL_CARET_RUN: horizontal caret run. @HB_OT_METRICS_TAG_HORIZONTAL_CARET_OFFSET: horizontal caret offset. @HB_OT_METRICS_TAG_VERTICAL_CARET_RISE: vertical caret rise. @HB_OT_METRICS_TAG_VERTICAL_CARET_RUN: vertical caret run. @HB_OT_METRICS_TAG_VERTICAL_CARET_OFFSET: vertical caret offset. @HB_OT_METRICS_TAG_X_HEIGHT: x height. @HB_OT_METRICS_TAG_CAP_HEIGHT: cap height. @HB_OT_METRICS_TAG_SUBSCRIPT_EM_X_SIZE: subscript em x size. @HB_OT_METRICS_TAG_SUBSCRIPT_EM_Y_SIZE: subscript em y size. @HB_OT_METRICS_TAG_SUBSCRIPT_EM_X_OFFSET: subscript em x offset. @HB_OT_METRICS_TAG_SUBSCRIPT_EM_Y_OFFSET: subscript em y offset. @HB_OT_METRICS_TAG_SUPERSCRIPT_EM_X_SIZE: superscript em x size. @HB_OT_METRICS_TAG_SUPERSCRIPT_EM_Y_SIZE: superscript em y size. @HB_OT_METRICS_TAG_SUPERSCRIPT_EM_X_OFFSET: superscript em x offset. @HB_OT_METRICS_TAG_SUPERSCRIPT_EM_Y_OFFSET: superscript em y offset. @HB_OT_METRICS_TAG_STRIKEOUT_SIZE: strikeout size. @HB_OT_METRICS_TAG_STRIKEOUT_OFFSET: strikeout offset. @HB_OT_METRICS_TAG_UNDERLINE_SIZE: underline size. @HB_OT_METRICS_TAG_UNDERLINE_OFFSET: underline offset.

hb_ot_name_id_t
hb_ot_name_id_t: @HB_OT_NAME_ID_INVALID: Value to represent a nonexistent name ID.

hb_ot_var_axis_flags_t
hb_ot_var_axis_flags_t: @HB_OT_VAR_AXIS_FLAG_HIDDEN: The axis should not be exposed directly in user interfaces.

hb_position_t

hb_reference_table_func_t

hb_script_t

hb_tag_t

hb_unicode_combining_class_func_t

hb_unicode_combining_class_t

hb_unicode_compose_func_t

hb_unicode_decompose_func_t

hb_unicode_general_category_func_t

hb_unicode_general_category_t

hb_unicode_mirroring_func_t

hb_unicode_script_func_t

hb_var_int_t

Unions

_hb_var_int_t